### 2015-06-02 #15 @eblondel
rsdmx package for R
- 1. What does this package do? (explain in 50 words or less)

`rsdmx` is a package to parse/read SDMX data and metadata documents in R. It provides a set of classes and methods to read data and metadata documents exchanged through the Statistical Data and Metadata Exchange (SDMX) framework. The package currently focuses on the SDMX XML standard format (SDMX-ML). It is part of the Web Technologies CRAN Task View. [More details](https://github.com/opensdmx/rsdmx/wiki#package_overview).
- 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

```
Package: rsdmx
Version: 0.4-7
Title: Tools for Reading SDMX Data and Metadata
  person("Emmanuel", "Blondel", role = c("aut", "cre"), email = "emmanuel.blondel1@gmail.com"),
  person("Matthieu", "Stigler", role = c("ctb")))
Maintainer: Emmanuel Blondel <emmanuel.blondel1@gmail.com>
Imports: methods, XML (>= 3.96-1.1), RCurl, plyr
Suggests: testthat
Description: Set of classes and methods to read data and metadata documents
  exchanged through the Statistical Data and Metadata Exchange (SDMX) framework,
  currently focusing on the SDMX XML standard format (SDMX-ML).
LazyLoad: yes
Collate:
  Class-SDMXSchema.R Class-SDMXType.R Class-SDMXStructureType.R
  Class-SDMXHeader.R Class-SDMXFooterMessage.R Class-SDMXFooter.R
  Class-SDMX.R Class-SDMXGenericData.R Class-SDMXCompactData.R Class-SDMXUtilityData.R
  Class-SDMXMessageGroup.R Class-SDMXConcept.R Class-SDMXConceptScheme.R Class-SDMXConcepts.R
  Class-SDMXCode.R Class-SDMXCodelist.R Class-SDMXCodelists.R Class-SDMXDimension.R
  Class-SDMXTimeDimension.R Class-SDMXPrimaryMeasure.R Class-SDMXAttribute.R Class-SDMXComponents.R
  Class-SDMXDataStructure.R Class-SDMXDataStructures.R Class-SDMXDataStructureDefinition.R
  SDMXSchema-methods.R SDMXType-methods.R SDMXStructureType-methods.R SDMXHeader-methods.R
  SDMXFooterMessage-methods.R SDMXFooter-methods.R SDMX-methods.R
  SDMXGenericData-methods.R SDMXCompactData-methods.R SDMXUtilityData-methods.R SDMXMessageGroup-methods.R
  SDMXConcept-methods.R SDMXConceptScheme-methods.R SDMXConcepts-methods.R SDMXCode-methods.R
  SDMXCodelist-methods.R SDMXCodelists-methods.R SDMXDimension-methods.R SDMXTimeDimension-methods.R
  SDMXPrimaryMeasure-methods.R SDMXAttribute-methods.R SDMXComponents-methods.R
  SDMXDataStructure-methods.R SDMXDataStructures-methods.R SDMXDataStructureDefinition-methods.R
  readSDMX.R
```
- 3. URL for the package (the development repository, not a stylized html page)

- 4. What data source(s) does it work with (if applicable)?

All SDMX-ML standard compliant data sources, supporting SDMX-ML format version `1.0`, `2.0` and `2.1`. Notice that a development strategy of `rsdmx` is to follow a low-level approach which intends to focus on the SDMX `format` specification, rather than the SDMX `web-service` specifications. `rsdmx` is then not restricted to the web (SDMX local files can be read) and for the web, it is not restricted to data sources disseminated with web-services compliant with the SDMX web-service specifications.
- 5. Who is the target audience?

Universities, Research scientists, statisticians, public institutions & companies that need to extract statistical data in R from institutions that disseminate data and metadata in the Statistical Data and Metadata eXchange format (SDMX).
- 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

As known active package, `RJSDMX` provides an interface to read SDMX data in R. AFAIK, the main differences i've identified are: (1) the core technology to read SDMX documents (`rsdmx` is writen in R and relies on the `XML` package, `RJSDMX` builds on a Java statistical software connector.), (2) `RJSDMX` restrains to SDMX data exchanged with known web-services strictly complying with the SDMX web-service standards, while `rsdmx` starts from a lower level flexible approach, focusing only on the SDMX _format_ specification (See above point 4.). As consequence, `rsdmx` does not bring (at now) helpers to "build" the SDMX web requests, but allows to read both local and remote SDMX data (from standard web-service architectures or not) and makes abstraction of the data providers and the data domain.
- 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
- [X] This package does not violate the Terms of Service of any service it interacts with.
- [X] The repository has continuous integration with Travis and/or another service
- [X] The package contains a vignette **(An online documentation is available and continuously updated. And i've added a quickstart vignette based on it.)**
- [X] The package contains a reasonably complete readme with devtools install instructions
- [X] The package contains unit tests
- [X] The package only exports functions to the NAMESPACE that are intended for end users
- 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
- [ ] Are there any package dependencies not on CRAN?
- [ ] Do you intend for this package to go on CRAN? **(yet on CRAN)**
- [X] Does the package have a CRAN accepted license?
- [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
- 9. Please add explanations below for any exceptions to the above:
- 10. If this is a resubmission following rejection, please explain the change in cirucmstances.
