### 2015-06-04 66cc6e4
#56 add SOAP response convenience handler
