### 2015-06-11 48681dc
#57 support for StructureSpecificData message type
