### 2015-06-14 #15 @jeroenooms
Thanks for this submission, Emmanuel!

I've been browsing through the code and documentation (which look good) but I am having difficulty wrapping my head around what the purpose of the sdmx format is, how it compares to other data formats, and how your package maps sdmx data to data frames in R.

I tried to compile the quickstart vignette but it gets stuck at the line

```r
stats <- as.data.frame(sdmx)
```

Also the vignette is still pretty technical and assumes the user is familiar with the sdmx format. I had never heard of this format before and the website isn't too helpful either so you'll have to help us a bit.

It would be very helpful to include a high-level vignette that introduces the main ideas from the sdmx format using worked examples with `rsdmx`. You don't have to discuss S4 classes or implementation details, those are sufficiently documented. The challenge is show some nice examples that illustrate in a few minutes what the sdmx format is used for and how your package makes it easy to access these data from R. Your readme lists a dozen of sdmx data sources, so it would be great if we can see the package in action on real-world data. Maybe you already have these examples and I just couldn't find them, in that case just put them in a nice vignette.

The jsonlite package has a few vignettes that illustrate the [mapping](http://cran.r-project.org/web/packages/jsonlite/vignettes/json-aaquickstart.html) and one with [web api examples](http://cran.r-project.org/web/packages/jsonlite/vignettes/json-apis.html). Something similar for sdmx would be very helpful to help the user get started with sdmx and rsdmx.
