### 2015-06-15 #15 @eblondel
Hello,

The SDMX format is a standard for data and metadata exchange between institutions. I'm not a SDMX data provider, so i cannot argue more for the SDMX format than other, but i know it is widely used by institutions that want to disseminate statistics, timeseries, whatever the domain. One benefit of SDMX is to bring with a dataset its complete data structure, named ``data structure definition`` in the SDMX vocabulary (metadata).

There is a good summary introduction to SDMX and their web-services [here](http://rbagd.eu/2015-03-28/downloading-data-with-sdmx-protocol-from-r.html), but this one explains more how to query SDMX datasets using the ``REST`` SDMX web-services, which then restrains to some data providers. One important aspect of ``rsdmx`` is that it doesn't restrain to the SDMX REST **web-services** specifications, but only focuses on the SDMX **format** specifications. The reason is that they are still many institutions that do not offer REST web-services, implement adhoc services or just provide SDMX files to download. Large data requests sometimes also redirect to adhoc download facilities. Hence, because ``rsdmx`` focuses on the format (format abstraction library), it allows reading data that it would not be possible to if it would be a pure ``REST`` service interface. ``rsdmx`` can then read local SDMX` data as well.

As explained in my email to ROpenSci, i had initially no vignette because preparing one bigger, as article, that explains the package in detail. For the purpose of this submission, i've finally decided to create one ``quickstart``. Nevertheless, this quickstart yet shows the package in action with real-world data flows. I've included examples with:
* **remote** datasets: If you feel it would add to the overall understanding, I might add to the quickstart a brief indication how to query SDMX web-services, but again this is out of scope of rsdmx which is a low-level library for SDMX, and not an interface to SDMX web-services.
* **local** datasets (after a bulk download of SDMX files)
* **metadata** documents (that describe a particular dataset)

Now i see the quickstart section of the README and it is the initial one.. I will update it from the quickstart vignette which provides various examples.

Regarding ``data.frame`` mappings, they are several ``as.data.frame`` ``S3`` methods (see [NAMESPACE](https://github.com/opensdmx/rsdmx/blob/master/NAMESPACE#L80)) according to the ``SDMX*`` ``S4`` object being manipulated. In SDMX, they are several data sub-formats, but generally two main one: ``Generic`` and ``Compact``, the latter is derivate in other formats specific to the data provider. Each sub-format brings a ``S3method`` to coerce into ``data.frame``.
For metadata, ``as.data.frame`` is available to read ``codelists`` and ``concepts`` (_ie_ reference data).
