+++ README.md
@@ -37,45 +37,116 @@ To unsubscribe, send an email to: [rsdmx+unsubscribe@googlegroups.com](rsdmx+uns
 
 ## Quickstart
 
-### Install
+[rsdmx](http://cran.r-project.org/web/packages/rsdmx/index.html) offers a low-level set of tools to read **data** and **metadata** in SDMX format. Its strategy is to make it very easy for the user. For this, a unique function named ``readSDMX`` has to be used, whatever it is a ``data`` or ``metadata`` document, or if it is ``local`` or ``remote`` datasource.
 
-Stable version from CRAN
 
-```r
+### Install rsdmx
+
+``rsdmx`` can be installed from CRAN
+```{r, echo = FALSE}
 install.packages("rsdmx")
 ```
-Or development version from GitHub.
 
-```r
+or from its development repository hosted in Github (using the ``devtools`` package):
+
+```{r, echo = FALSE}
 devtools::install_github("opensdmx/rsdmx")
 ```
 
-```r
-library("rsdmx")
+### Load rsdmx
+
+To load rsdmx in R, do the following:
+
+```{r, echo = FALSE}
+library(rsdmx)
 ```
 
-### Use cases
+### Read dataset documents
 
-#### Read SDMX-ML **data** documents
+This section will introduce you on how to read SDMX *dataset* documents, either from remote datasources, or from local SDMX files.
 
-```r
-dataURL <- "http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011"
-sdmx <- readSDMX(dataURL)
+#### Read _remote_ datasets
+
+The following code snipet shows you how to read a dataset from a remote data source, taking as example the [FAO data portal](http://data.fao.org/sdmx/index.html): [http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008](http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008)
+
+```{r, echo = FALSE}
+myUrl <- "http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008"
+dataset <- readSDMX(myUrl)
+stats <- as.data.frame(dataset) 
+```
+
+You can try it out with other datasources, such as:
+* [**OECD StatExtracts portal**](http://stats.oecd.org): [http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011](http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011)
+* [**EUROSTAT portal**](http://epp.eurostat.ec.europa.eu/portal/page/portal/sdmx_web_services/getting_started/rest_sdmx_2.1): [http://ec.europa.eu/eurostat/SDMX/diss-web/rest/data/cdh_e_fos/..PC.FOS1.BE/?startperiod=2005&endPeriod=2011](http://ec.europa.eu/eurostat/SDMX/diss-web/rest/data/cdh_e_fos/..PC.FOS1.BE/?startperiod=2005&endPeriod=2011)
+* [**European Central Bank (ECB)**](https://sdw-wsrest.ecb.europa.eu): [https://sdw-wsrest.ecb.europa.eu/service/data/DD/M.SE.BSI_STF.RO.4F_N](https://sdw-wsrest.ecb.europa.eu/service/data/DD/M.SE.BSI_STF.RO.4F_N)
+
[…]
