### 2015-06-18 31179ab
Update quickstart.Rmd
