### 2015-06-18 #15 @eblondel
@jeroenooms
* I've slightly improved the quickstart vignette https://github.com/opensdmx/rsdmx/blob/master/vignettes/quickstart.Rmd I've added a quick introduction to SDMX essentials, in order to familiarize with SDMX before jumping into the R part.
* i've also explained a bit more what ``rsdmx`` covers in term of functionality (focus on SDMX format vs. web-services API).
* i've also added a link to a short intro on SDMX which extends the quick intro provided in the vignette.
Hope this helps
