### 2015-06-18 #15 @jeroenooms
@eblondel I tried building the vignette and I get this: http://rpubs.com/jeroenooms/rsdmx-quickstart. Is that the intended output?
