### 2015-06-18 8168cf4
fix malformed elements of the vignette
