+++ vignettes/quickstart.Rmd
@@ -12,12 +12,14 @@ The goal of this document is to get you up and running with rsdmx as quickly as 
 ## SDMX - a short introduction
 
 The SDMX framework provides two sets of standard specifications to facilitate the exchange of statistical data:
-(1) standard formats
-(2) web-service specifications
+
+* standard formats
+* web-service specifications
 
 SDMX allows to disseminate both **data** (a dataset) and **metadata** (the description of the dataset).
 
 For this, the SDMX standard provides various types of _documents_, also known as _messages_. Hence there will be:
+
 * **data** SDMX-ML _documents_. The two main _document_ types are the ``Generic`` and ``Compact`` ones. The latter aims to provide a more compact XML document. They are other data _document_ derivating from the main types.
 * **metadata** SDMX-ML _documents_. The main metadata _document_ is known a ``Data Structure Definition`` (DSD). As its name indicates, it _describes_ the structure and organization of a dataset, and will generally include all the master/reference data used to characterize a dataset. The 2 main types of metadata are (1) the ``concepts``, which correspond the _dimensions_ and/or _attributes_ of the dataset, and (2) the ``codelists`` which inventory the possible values to be used in the representation of _dimensions_ and _attributes_.
 
@@ -34,14 +36,9 @@ Let's see then how to use ``rsdmx``!
 
 ## Install rsdmx
 
-``rsdmx`` can be installed from CRAN
-```{r, echo = FALSE}
-install.packages("rsdmx")
-```
+``rsdmx`` can be installed from CRAN or from its development repository hosted in Github (using the ``devtools`` package). For the latter, you will need the ``devtools`` package and run:
 
-or from its development repository hosted in Github (using the ``devtools`` package):
-
-```{r, echo = FALSE}
+```{r, eval=FALSE}
 devtools::install_github("opensdmx/rsdmx")
 ```
 
@@ -49,7 +46,7 @@ devtools::install_github("opensdmx/rsdmx")
 
 To load rsdmx in R, do the following:
 
-```{r, echo = FALSE}
+```{r, eval=FALSE}
 library(rsdmx)
 ```
 
@@ -57,28 +54,29 @@ library(rsdmx)
 
 This section will introduce you on how to read SDMX *dataset* documents, either from remote datasources, or from local SDMX files.
 
-##### Read _remote_ datasets
+### Read _remote_ datasets
 
 The following code snipet shows you how to read a dataset from a remote data source, taking as example the [FAO data portal](http://data.fao.org/sdmx/index.html): [http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008](http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008)
 
-```{r, echo = FALSE}
+```{r, eval=FALSE}
 myUrl <- "http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008"
 dataset <- readSDMX(myUrl)
 stats <- as.data.frame(dataset) 
 ```
 
 You can try it out with other datasources, such as:
+
[…]
