### 2015-06-18 #15 @eblondel
@jeroenooms i've fixed the vignette. Thanks
