### 2015-06-18 #15 @jeroenooms
Ok so you added `eval=FALSE` but the examples is still broken? It gives a ton of warnings...
