### 2015-06-18 #15 @eblondel
I've added ``eval=FALSE`` indeed. Examples are NOT broken, they work. They would be broken with an ``error`` and a wrong result, not with a ``warning``. Anyway, i'm currently looking into the first set of warnings (only) to fix it since these particular warnings are unexpected, but it's not the case of all warnings. Is there  any way in RMarkdown to ``eval`` an example, and switch off ``warning`` (without calling ``invisible`` for sure)? I keep you update.
