### 2015-06-18 ae7a100
suppress warnings for rbind in case occasional obs status
