### 2015-06-18 #15 @eblondel
@jeroenooms I've further tested the vignette with RStudio. As said, the ``warnings`` were not blocking, and examples are working. I've fixed the warnings anyway which i agree were annoying. I've let ``eval=FALSE`` for the ``install_github`` code (so ensure you update the package before running the vignette). Everything should be ok now.
