### 2015-06-18 #15 @jeroenooms
Those warnings usually indicate that there are bugs in concatenating vectors which might lead to the wrong data being inserted in the dataframe. I would much prefer if you try to find the origin of these warnings rather than muffle things under the carpet with `suppressWarnings`...
