### 2015-06-18 #15 @eblondel
I'm not "muffling things under the carpet", and if i've put a ``suppressWarnings``, it is because i found the origin of these warnings: observation status are not mandatory in SDMX, so some observations have, some other do not, I would not put a ``suppressWarnings`` if i would have a doubt or would not know the origin of it.
