### 2015-06-19 #15 @eblondel
Alternatively, i might use ``rbind.fill`` from ``plyr`` package (because it is yet a declared dependency), but IHMO it is not necessary here, a ``suppressWarnings`` is enough to deal with this case for which the origin is well-known.
