### 2015-06-19 #15 @jeroenooms
This warning always makes me very nervous, it is rarely ignorable. It means that you are combining data of incompatible dimensions and R is repeating a vector to fill things up. It would be much better practice if you can construct your data frame in such a way that it does not trigger this warning in the first place, rather than suppressing it.
