### 2015-06-19 #15 @eblondel
I understand your point. When i say that the origin is well-identified is because it deals with the standard SDMX-ML spec itself. In the compact data format (this what we are talking about), an ``observation`` comes with attributes, some are mandatory, some optional. This means that according to the ``observation`` you will not have the same number of ``XML`` attributes. One key point is that there is no clear statement about the naming of these attributes in the standard, which means that i cannot pre-build a ``data.frame`` looking for particular attribute ``names`` that are not known a priori. As said, an alternative not to use ``suppressWarnings`` is to use ``rbind.fill``.
