### 2015-06-23 #15 @jeroenooms
@eblondel do you happen to be at the french r meeting in grenoble this week?
