### 2015-06-23 #15 @eblondel
Hum, no, i was not aware of it, i'm currently on duty travel until next week. Are you going to attend?
