### 2015-06-23 #15 @jeroenooms
Yes I'm in Grenoble all week. Will you attend useR 2015 in Denmark?
