### 2015-06-23 #15 @eblondel
No, unfortunately :-(, i could attend to useR 2015 in Denmark. I know that some people (Rytis Bagdziunas) will talk about ``rsdmx`` in his poster named "Short-term forecasting with factor models".
