### 2015-06-23 #15 @jeroenooms
Okay thats good to know, thank you. Do you know where I can find the `dynfactoR` package? Perhaps that will help me understand better how `rsdmx` is used in practice.
