### 2015-06-23 #15 @eblondel
The author is on [Github](https://github.com/rbagd?tab=repositories), but it seems ``dynfactorR`` is not available online (at least at now). You may contact him directly. Alternatively, you can also have a look to its repositories, where he put in place some use cases based on ``rsdmx``. See this one for example: [https://github.com/rbagd/eu-unemployment](https://github.com/rbagd/eu-unemployment) Hope this helps
