### 2015-07-06 #15 @eblondel
@jeroenooms If it helps, you can see https://github.com/rbagd/useR2015 (just found it, it is the poster i mentioned in some past comment).

I have also been informed of a nice use case using ``opencpu`` to build a SDMX viewer, explained in this post: https://medium.com/@ggruic/what-is-open-data-and-how-can-we-access-it-d47f0089e330 (see https://github.com/ggruic/sdmxquickview).
