### 2015-07-06 #15 @jeroenooms
I'll return from traveling tomorrow, then I'll have another look this week. Also I saw a poster at useR about  the `RJSDMX` package. Maybe you can discuss the differences with this package a bit as well.
