### 2015-07-06 5b6f91d
improve obs attrs handling for compact data
