### 2015-07-06 #15 @eblondel
Thanks, as follow-up, i have improved a bit the code related to what we discussed (``warnings`` raised for compact data observation attributes). https://github.com/opensdmx/rsdmx/commit/5b6f91d37c312f3e5cc88588fb9ac92e99c1d075

For what concerns the main differences with RJSDMX, i've summarized them in this ticket description (see point 6). As initially explained to ropensci, I intend to add in the future another more technical vignette but in the form of paper, and it is planned to discuss rsdmx by comparing it further with other packages.
