### 2015-07-08 f2bf83f
#58 add control to find/remove UTF-8 BOM
