### 2015-07-21 1f52cb2
add CRAN badge
