+++ README.md
@@ -1,6 +1,7 @@
 **rsdmx**
 =======
 [![Build Status](https://travis-ci.org/opensdmx/rsdmx.svg?branch=master)](https://travis-ci.org/opensdmx/rsdmx.svg?branch=master)
+[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/rsdmx)](http://cran.r-project.org/web/packages/rsdmx)
 
 ```rsdmx``: Tools  for reading SDMX data and metadata documents in R
 
