+++ .travis.yml
@@ -9,13 +9,14 @@ before_install:
   - curl -OL http://raw.github.com/craigcitro/r-travis/master/scripts/travis-tool.sh
   - chmod 755 ./travis-tool.sh
   - ./travis-tool.sh bootstrap
+  - pip install codecov
 install:
   - ./travis-tool.sh install_deps
 script: ./travis-tool.sh run_tests
-
+after_success:
+  - codecov
 after_failure:
   - ./travis-tool.sh dump_logs
-
 notifications:
   email:
     on_success: change
