+++ .travis.yml
@@ -9,12 +9,13 @@ before_install:
   - curl -OL http://raw.github.com/craigcitro/r-travis/master/scripts/travis-tool.sh
   - chmod 755 ./travis-tool.sh
   - ./travis-tool.sh bootstrap
-  - pip install codecov
 install:
   - ./travis-tool.sh install_deps
 script: ./travis-tool.sh run_tests
+r_github_packages:
+  - jimhester/covr
 after_success:
-  - codecov
+  - Rscript -e 'covr::codecov()'
 after_failure:
   - ./travis-tool.sh dump_logs
 notifications:
