### 2015-07-21 649e7cd
adding code coverage
