### 2015-07-23 c6a70e7
Try to require the library
