+++ .travis.yml
@@ -15,7 +15,7 @@ script: ./travis-tool.sh run_tests
 r_github_packages:
   - jimhester/covr
 after_success:
-  - Rscript -e 'covr::codecov()'
+  - Rscript -e 'library(covr); codecov()'
 after_failure:
   - ./travis-tool.sh dump_logs
 notifications:
