### 2015-07-23 83b5bba
try directly installing
I think C does not install from r_github_packages
