### 2015-08-27 b76b993
Add vignetteBuilder
