### 2015-08-27 bd9306a
Add knitr in 'Suggests'
