+++ DESCRIPTION
@@ -8,14 +8,13 @@ Authors@R: c(
 Maintainer: Emmanuel Blondel <emmanuel.blondel1@gmail.com>
 Depends: R (>= 2.15)
 Imports: methods, XML (>= 3.96-1.1), RCurl, plyr, utils
-Suggests: testthat, knitr
+Suggests: testthat
 Description: Set of classes and methods to read data and metadata documents
   exchanged through the Statistical Data and Metadata Exchange (SDMX) framework,
   currently focusing on the SDMX XML standard format (SDMX-ML).
 License: GPL (>= 2)
 URL: https://github.com/opensdmx/rsdmx, http://www.sdmx.org
 BugReports: https://github.com/opensdmx/rsdmx/issues
-VignetteBuilder: knitr
 LazyLoad: yes
 Collate:
   Class-SDMXSchema.R Class-SDMXType.R Class-SDMXStructureType.R
