### 2015-09-15 9679252
#63 DataStructuresType is now a valid structure type
