### 2015-09-15 cceadcb
#64 make SDMXComponents optional
