### 2015-09-15 86d11ba
#65 support for DataStructures as.data.frame method
