+++ R/SDMXDataStructures-methods.R
@@ -39,3 +39,52 @@ datastructures.SDMXDataStructures <- function(xmlObj){
   return(datastructures)
 }
 
+#methods
+as.data.frame.SDMXDataStructures <- function(x, ...){
+  
+  out <- do.call("rbind.fill",
+                 lapply(x@datastructures,
+                        function(ds){
+                          
+                          names <- slot(ds, "Name")
+                          dsf.names <- as.data.frame(names, stringsAsFactors = FALSE)
+                          colnames(dsf.names) <- paste0("Name.", colnames(dsf.names))
+                          
+                          desc <- slot(ds, "Description")
+                          dsf.desc <- NULL
+                          if(length(desc) > 0){
+                            dsf.desc <- as.data.frame(desc, stringsAsFactors = FALSE)
+                            colnames(dsf.desc) <- paste0("Description.", colnames(dsf.desc))
+                          }
+                          
+                          dsf <- data.frame(
+                              id = slot(ds, "id"),
+                              agencyID = slot(ds, "agencyID"),
+                              dsf.names,
+                              stringsAsFactors = FALSE)
+                          
+                          if(!is.null(dsf.desc)){
+                            dsf <- cbind(dsf, dsf.desc, stringsASFactors = FALSE)
+                          }
+                           
+                          dsf <- cbind(dsf,
+                                       version = slot(ds, "version"),
+                                       uri = slot(ds, "uri"),
+                                       urn = slot(ds, "urn"),
+                                       isExternalReference = slot(ds, "isExternalReference"),
+                                       isFinal = slot(ds, "isFinal"),
+                                       validFrom = slot(ds, "validFrom"),
+                                       validTo = slot(ds, "validTo"),
+                                       stringsAsFactors = FALSE
+                                       )
+                              
+                          
+                          return(dsf)
+                        })
+                 )
+  return(out)
+  
+}
+
+setAs("SDMXDataStructures", "data.frame",
+      function(from) as.data.frame.SDMXDataStructures(from));
