+++ man/rsdmx-package.Rd
@@ -15,8 +15,8 @@
 Package: \tab rsdmx\cr
 Type: \tab Package\cr
 Version
-: \tab 0.4-7\cr
-Date: \tab 2015-07-14\cr
+: \tab 0.5-0\cr
+Date: \tab 2015-09-15\cr
 License: \tab GPL(>=2.0)\cr
 LazyLoad: \tab yes\cr
 }
