+++ NAMESPACE
@@ -85,3 +85,4 @@ S3method(as.data.frame, SDMXStructureSpecificData)
 S3method(as.data.frame, SDMXMessageGroup)
 S3method(as.data.frame, SDMXConcepts)
 S3method(as.data.frame, SDMXCodelists)
+S3method(as.data.frame, SDMXDataStructures)
