### 2015-09-15 30f58b9
#65 fix unit test
