### 2015-09-15 aec10f8
#65 add example on data structures
