### 2015-09-16 3153c06
Add badge to give rsdmx version on Github master
