+++ R/readSDMX.R
@@ -1,105 +1,143 @@
 # E.Blondel - 2013/06/10
 #=======================
 
-# function to read SDMX as character string
+# function to read SDMX as character string or from helper arguments
 # (required in order to encapsulate a S3 old class object in a S4 representation)
-readSDMX <- function(file, isURL = TRUE){
-	
-	#load data
-	status <- 0
-	if(isURL == FALSE){
-		if(!file.exists(file))
-			stop("File ", file, "not found\n")
-		xmlObj <- xmlTreeParse(file, useInternalNodes = TRUE)
-    		status <- 1
-	}else{
-		rsdmxAgent <- paste("rsdmx/",as.character(packageVersion("rsdmx")),sep="")
-		content <- getURL(file, httpheader = list('User-Agent' = rsdmxAgent),
+readSDMX <- function(file = NULL, isURL = TRUE,
+                     provider = NULL, id = NULL,
+                     operation = NULL, key = NULL, filter = NULL, start = NULL, end = NULL) {
+  
+  #check from arguments if request has to be performed
+  buildRequest <- FALSE
+  if(!is.null(provider)){
+    if(class(provider) != "SDMXServiceProvider"){
+      stop("Provider should be an instance of 'SDMXServiceProvider'")
+    }
+    buildRequest <- TRUE
+  }
+  
+  if(!is.null(id)){
+    provider <- findSDMXServiceProvider(id)
+    if(is.null(provider)){
+      stop("No provider with identifier ", id)
+    }
+    provider <- provider[[1]]
+    buildRequest <- TRUE
+  }
+  
+  #proceed with the request build
+  if(buildRequest){
+    if(is.null(operation)) stop("SDMX service operation cannot be null")
+    if(is.null(key)) stop("SDMX service key cannot be null")
+    file <- provider@builder@handler(operation = operation,
+                             key = key,
+                             filter = filter,
+                             start = start,
+                             end = end)
+    message(file)
+  }
+  
+  #call readSDMX original
+  if(is.null(file)) stop("Empty file argument")
+  if(buildRequest) isURL = TRUE
+  
+  #load data
+  status <- 0
+  if(isURL == FALSE){
+    if(!file.exists(file))
+      stop("File ", file, "not found\n")
+    xmlObj <- xmlTreeParse(file, useInternalNodes = TRUE)
+    status <- 1
+  }else{
+    rsdmxAgent <- paste("rsdmx/",as.character(packageVersion("rsdmx")),sep="")
+    content <- getURL(file, httpheader = list('User-Agent' = rsdmxAgent),
                       ssl.verifypeer = FALSE, .encoding = "UTF-8")
     
-		status <- tryCatch({
-			if(attr(regexpr("<!DOCTYPE html>", content), "match.length") == -1){
+    status <- tryCatch({
+      if(attr(regexpr("<!DOCTYPE html>", content), "match.length") == -1){
         
-			        #check the presence of a BOM
-			        BOM <- "\ufeff"
-			        if(attr(regexpr(BOM, content), "match.length") != - 1){
-			          content <- gsub(BOM, "", content)
-			        }
+        #check the presence of a BOM
+        BOM <- "\ufeff"
+        if(attr(regexpr(BOM, content), "match.length") != - 1){
+          content <- gsub(BOM, "", content)
+        }
         
-			        xmlObj <- xmlTreeParse(content, useInternalNodes = TRUE)
-			        status <- 1
-			}else{
-				stop("The SDMX web-request failed. Please retry")
-		      	}
-		},error = function(err){
-			print(err)
-			status <<- 0
-			return(status)
-		})
-	}
[…]
