+++ R/profile.R
@@ -1,0 +1,4 @@
+.onLoad <- function (libname, pkgname) {
+  assign(".rsdmx.options", new.env(), envir= asNamespace(pkgname))
+  setSDMXServiceProviders()
+}
