+++ tests/testthat/test_Main_Helpers.R
@@ -1,0 +1,52 @@
+# test_Main_Helpers.R
+# Author: Emmanuel Blondel <emmanuel.blondel1@gmail.com>
+#
+# Description: Unit tests for SDMX Main methods
+# using helpers to build the SDMX request
+#=======================
+require(rsdmx, quietly = TRUE)
+require(testthat)
+context("SDMX")
+
+#SDMX datasets
+#------------
+
+#ECB
+#test_that("ECB",{
+#  sdmx <- readSDMX(id = "ECB", operation = "data",
+#                   key = "DD", filter = "M.SE.BSI_STF.RO.4F_N",
+#                   start = 2010, end = 2010)
+#  expect_is(sdmx, "SDMXGenericData")
+#})
+
+#EUROSTAT
+test_that("EUROSTAT",{
+  sdmx <- readSDMX(id = "EUROSTAT", operation = "data",
+                   key = "cdh_e_fos", filter = "..PC.FOS1.BE",
+                   start = 2005, end = 2010)
+  expect_is(sdmx, "SDMXGenericData")
+})
+
+#OECD
+test_that("OECD",{
+  sdmx <- readSDMX(id = "OECD", operation = "GetData",
+                   key = "MIG", filter = "TOT..", start = 2011, end = 2011)
+  expect_is(sdmx, "SDMXMessageGroup")
+})
+
+#UN-FAO
+test_that("UN-FAO",{
+  sdmx <- readSDMX(id = "UN-FAO", operation = "data",
+                   key = "CROP_PRODUCTION", filter = ".156.5312..",
+                   start = "2010", end = "2014")
+  expect_is(sdmx, "SDMXGenericData")
+})
+
+#UN-ILO
+test_that("UN-ILO",{
+  sdmx <- readSDMX(id = "UN-ILO", operation = "data",
+                   key = "DF_CPI_FRA_CPI_TCPI_COI_RT", filter = "ALL",
+                   start = "2014-01-01", end = "2014-12-31")
+  expect_is(sdmx, "SDMXGenericData")
+})
+

[…]
