### 2015-09-22 9772c38
#45 fix some docs and add more tests
