### 2015-09-23 8a5729e
#45 add some online documentation
