+++ vignettes/quickstart.Rmd
@@ -73,6 +73,15 @@ You can try it out with other datasources, such as:
 
 The online rsdmx documentation also provides a list of data providers, either from international or national institutions.
 
+Now, the service providers above mentioned are known by ``rsdmx`` which let users using ``readSDMX`` with the helper parameters. Let's see how it would look like for querying an OECD datasource:
+
+```{r, echo = FALSE}
+sdmx <- readSDMX(id = "OECD", operation = "GetData", key = "MIG", filter = "TOT..", start = 2011, end = 2011)
+df <- as.data.frame(sdmx)
+head(df)
+```
+
+
 ### Read _local_ datasets
 
 This example shows you how to use ``rsdmx`` with _local_ SDMX files, previously downloaded from [EUROSTAT](http://ec.europa.eu/eurostat).
