+++ README.md
@@ -42,6 +42,8 @@ To unsubscribe, send an email to: [rsdmx+unsubscribe@googlegroups.com](rsdmx+uns
 
 [rsdmx](http://cran.r-project.org/package=rsdmx) offers a low-level set of tools to read **data** and **metadata** in SDMX format. Its strategy is to make it very easy for the user. For this, a unique function named ``readSDMX`` has to be used, whatever it is a ``data`` or ``metadata`` document, or if it is ``local`` or ``remote`` datasource.
 
+It is important to highlight that one of the major benefits of ``rsdmx`` is to focus first on the SDMX **format** specifications (acting as format abstraction library). This allows ``rsdmx`` reading SDMX data from _remote_ datasources, or from _local_ SDMX files. For accessing _remote_ datasources, it also means that ``rsdmx`` does not bound to SDMX **service** specifications, and can read a wider ranger of datasources.
+
 
 ### Install rsdmx
 
@@ -64,11 +66,110 @@ To load rsdmx in R, do the following:
 library(rsdmx)
 ```
 
+### readSDMX & helper functions
+
+
+#### readSDMX as low-level function
+
+The ``readSDMX`` function is then first designed at low-level so it can take as parameters a _url_ (``isURL=TRUE`` by default) or a _file_. So wherever is located the SDMX document, ``readSDMX`` will allow you to read it, as follows:
+
+```{r, echo = FALSE}
+
+  #read a remote file
+  sdmx <- readSDMX(file = "someUrl")
+  
+  #read a local file
+  sdmx <- readSDMX(file = "somelocalfile", isURL = FALSE)
+
+```
+
+In addition, in order to facilitate querying datasources, ``readSDMX`` also providers helpers to query well-known remote datasources. This allows not to specify the entire URL, but rather specify a simple provider ID, and the different parameters to build a SDMX query (e.g. for a dataset query: operation, key, filter, startPeriod and endPeriod). 
+
+This is made possible as a list of SDMX service providers is embedded within ``rsdmx``, and such list provides all the information required for ``readSDMX`` to build the SDMX request (url) before accessing the datasource.
+
+
+#### get list of SDMX service providers
+
+The list of known SDMX service providers can be queried as follows:
+
+```{r, echo = FALSE}
+
+providers <- getSDMXServiceProviders()
+
+#list all provider ids
+sapply(providers, function(x) slot(x, "id"))
+
+```
+
+#### create/add a SDMX service provider
+
+It also also possible to create and add a new SDMX service providers in this list (so ``readSDMX`` can be aware of it). A provider can be created with the ``SDMXServiceProvider``, and is made of three parameters: an ``id``, its ``name``, and a request ``builder``.
+
+The request builder can be created with ``SDMXRequestBuilder`` which takes 3 arguments: the ``baseUrl`` of the service endpoint, a service url ``suffix``, and a ``handler`` function which will allow to build the web request.
+
+``rsdmx`` intends to provider specific request builder that embedds yet an handler function (not need to implement it), and is now attempting to provide a ``SDMX21RequestBuilder`` to build SDMX 2.1 REST web-requests. All this is still under experiments.
+
+Let's see it with an example:
+
+First create a request builder for our provider:
+
+```{r, echo = FALSE}
+
+myBuilder <- SDMXRequestBuilder(
+  baseUrl = "http://www.myorg.org/sdmx",
+  suffix = "service",
+  handler = function(baseUrl, operation, key, filter, suffix, start, end){
[…]
