### 2015-09-24 f666fc4
#45 enable native R filtering + enrich readme
