+++ tests/testthat/test_Main_Helpers.R
@@ -22,7 +22,7 @@ context("SDMX")
 #EUROSTAT
 test_that("EUROSTAT",{
   sdmx <- readSDMX(id = "EUROSTAT", operation = "data",
-                   key = "cdh_e_fos", filter = "..PC.FOS1.BE",
+                   key = "cdh_e_fos", filter = list(NULL, NULL, "PC", "FOS1", "BE"),
                    start = 2005, end = 2010)
   expect_is(sdmx, "SDMXGenericData")
 })
@@ -30,14 +30,14 @@ test_that("EUROSTAT",{
 #OECD
 test_that("OECD",{
   sdmx <- readSDMX(id = "OECD", operation = "GetData",
-                   key = "MIG", filter = "TOT..", start = 2011, end = 2011)
+                   key = "MIG", filter = list("TOT", NULL, NULL), start = 2011, end = 2011)
   expect_is(sdmx, "SDMXMessageGroup")
 })
 
 #UN-FAO
 test_that("UN-FAO",{
   sdmx <- readSDMX(id = "UN-FAO", operation = "data",
-                   key = "CROP_PRODUCTION", filter = ".156.5312..",
+                   key = "CROP_PRODUCTION", filter = list(NULL, "156", "5312", NULL, NULL),
                    start = "2010", end = "2014")
   expect_is(sdmx, "SDMXGenericData")
 })
@@ -45,7 +45,7 @@ test_that("UN-FAO",{
 #UN-ILO
 test_that("UN-ILO",{
   sdmx <- readSDMX(id = "UN-ILO", operation = "data",
-                   key = "DF_CPI_FRA_CPI_TCPI_COI_RT", filter = "ALL",
+                   key = "DF_CPI_FRA_CPI_TCPI_COI_RT", filter = "ALL", filter.native = FALSE,
                    start = "2014-01-01", end = "2014-12-31")
   expect_is(sdmx, "SDMXGenericData")
 })
