### 2015-09-24 7699c60
#45 further improvements and fixs
