+++ README.md
@@ -106,17 +106,17 @@ The list of known SDMX service providers can be queried as follows:
 providers <- getSDMXServiceProviders()
 
 #list all provider ids
-sapply(providers, function(x) slot(x, "id"))
+sapply(providers, function(x) slot(x, "agencyId"))
 
 ```
 
 #### create/add a SDMX service provider
 
-It also also possible to create and add a new SDMX service providers in this list (so ``readSDMX`` can be aware of it). A provider can be created with the ``SDMXServiceProvider``, and is made of three parameters: an ``id``, its ``name``, and a request ``builder``.
+It also also possible to create and add a new SDMX service providers in this list (so ``readSDMX`` can be aware of it). A provider can be created with the ``SDMXServiceProvider``, and is made of three parameters: an ``agencyId``, its ``name``, and a request ``builder``.
 
-The request builder can be created with ``SDMXRequestBuilder`` which takes 3 arguments: the ``baseUrl`` of the service endpoint, a service url ``suffix``, and a ``handler`` function which will allow to build the web request.
+The request builder can be created with ``SDMXRequestBuilder`` which takes 3 arguments: the ``baseUrl`` of the service endpoint, a ``suffix`` logical parameter (either the ``agencyId`` has to be used as suffix in the web-request), and a ``handler`` function which will allow to build the web request.
 
-``rsdmx`` intends to provider specific request builder that embedds yet an handler function (not need to implement it), and is now attempting to provide a ``SDMX21RequestBuilder`` to build SDMX 2.1 REST web-requests. All this is still under experiments.
+``rsdmx`` intends to provider specific request builder that embedds yet an handler function (not need to implement it), and is now attempting to provide a ``SDMXRESTRequestBuilder`` to build SDMX REST web-requests. All this is still under experiments.
 
 Let's see it with an example:
 
@@ -126,14 +126,14 @@ First create a request builder for our provider:
 
 myBuilder <- SDMXRequestBuilder(
   baseUrl = "http://www.myorg.org/sdmx",
-  suffix = "service",
-  handler = function(baseUrl, operation, key, filter, suffix, start, end){
-    paste(baseUrl, operation, key, filter, paste0(suffix, "?startPeriod=", start, "&endPeriod=", end), sep="/")
+  suffix = TRUE,
+  handler = function(baseUrl, agencyId, suffix, operation, key, filter, suffix, start, end){
+    paste(baseUrl, operation, key, filter, paste0(agencyId, "?startPeriod=", start, "&endPeriod=", end), sep="/")
   }
 )
 ```
 
-As you can see, we built a handler that will be in charge of creating a web-request such as [http://www.myorg.org/sdmx/operation/key/filter/suffix?startPeriod=start&endPeriod=end](http://www.myorg.org/sdmx/operation/key/filter/suffix?startPeriod=start&endPeriod=end)
+As you can see, we built a handler that will be in charge of creating a web-request such as [http://www.myorg.org/sdmx/operation/key/filter/agencyId?startPeriod=start&endPeriod=end](http://www.myorg.org/sdmx/operation/key/filter/agencyId?startPeriod=start&endPeriod=end)
 
 We can create a provider with the above request builder, and add it to the list of known SDMX service providers:
 
@@ -141,7 +141,7 @@ We can create a provider with the above request builder, and add it to the list 
 
 #create the provider
 provider <- SDMXServiceProvider(
-id = "MYORG",
+agencyId = "MYORG",
 name = "My Organization",
 builder = myBuilder
 )
@@ -150,7 +150,7 @@ builder = myBuilder
 addSDMXServiceProvider(provider)
 
[…]
