+++ doc/rsdmx_presentation.tex
@@ -1,0 +1,611 @@
+\documentclass{beamer}
+  \usepackage[utf8]{inputenc}
+  \usepackage{hyperref}
+  \usepackage{xspace}
+  \usepackage{xcolor}
+  \usepackage{graphicx}
+  \usepackage{Sweave}
+  \usepackage{multicol}
+  
+  \definecolor{links}{HTML}{2A1B81}
+  \hypersetup{colorlinks,linkcolor=,urlcolor=links}  
+  
+  %Beamer properties
+  \usetheme{Warsaw} %{Boadilla}
+  \useoutertheme{miniframes}
+  \usecolortheme{seahorse}
+  \usecolortheme{rose}
+  \setbeamerfont{footnote}{size=\tiny}
+
+  % Add R Logo
+  \definecolor{Rcolor}{RGB}{150,160,190}
+
+  \newcommand{\Rx}{\fontsize{10pt}{12pt}\selectfont
+  	\raisebox{.3em}{\hspace{1.2em}%
+  		\llap{\resizebox{1.09em}{.5em}{\color{black}$\bigcirc$}}%
+  		\llap{\resizebox{1.199em}{.55em}{\color{darkgray}$\bigcirc$}}%
+  		\llap{\resizebox{1.19em}{.52em}{\color{gray!50}$\bigcirc$}}%
+		\llap{\resizebox{1.1em}{.5em}{\color{gray}$\bigcirc$}}%
+		\llap{\resizebox{1.25em}{.55em}{\color{gray}$\bigcirc$}}%
+	}%
+	\hspace{-.85em}%
+	\textbf{%
+		%\resizebox{.55em}{1.5ex}{\textcolor{black!80}{\textsf{R}}}%
+		\textcolor{black}{\textsf{R}}%
+		\hspace{-.025em}\raisebox{.01em}{\llap{\textcolor{Rcolor}{\textsf{R}}}}%
+	}}%
+
+  \newbox\rbox
+  \savebox\rbox{\scalebox{0.1}{\Rx}}
+  \makeatletter
+  \def\R{\scalebox{\f@size}{\usebox\rbox}\xspace}
+  \makeatletter
+  
+  %S4 representations
+  \newenvironment<>{rsdmx}[1]{%
+  \DefineVerbatimEnvironment{Soutput}{Verbatim}{fontsize=\tiny,formatcom=\color{black}}
+  \setbeamerfont{block title}{size=\normalsize}
+  \setbeamerfont{block body}{size=\scriptsize}
+  \setbeamercolor{block title}{fg=white,bg=gray!90!black}%
+  \begin{block}#2{#1 \R representation}}{\end{block}}
+  
+  %R code
+  \DefineVerbatimEnvironment{Sinput}{Verbatim}{fontsize=\scriptsize,formatcom=\color{blue}}
+  \DefineVerbatimEnvironment{Soutput}{Verbatim}{fontsize=\tiny,formatcom=\color{gray}}
+
+  %RSDMX Presentation%
+  %------------------%
+
+  \title{rsdmx - Tools for reading SDMX data and metadata documents in \R}
+  \author[E. Blondel]{Emmanuel Blondel \\ \texttt{emmanuel.blondel1@gmail.com}}
+  \institute{CEO - International Consultant}
+  \date{November 12, 2014}
+
+\AtBeginSection[]
+{
+  \begin{frame}<beamer>
+    \frametitle{Outline}
+    \tableofcontents[currentsection]
+  \end{frame}
+}
+
+%\setbeamertemplate{section in toc}{\protect\block{Research Question~\inserttocsectionnumber}\inserttocsection\protect\endblock}
+
+\begin{document}
+
+  \begin{frame}
[…]
