+++ doc/rsdmx_presentation.tex
@@ -59,7 +59,7 @@
   \title{rsdmx - Tools for reading SDMX data and metadata documents in \R}
   \author[E. Blondel]{Emmanuel Blondel \\ \texttt{emmanuel.blondel1@gmail.com}}
   \institute{CEO - International Consultant}
-  \date{November 12, 2014}
+  \date{September 27, 2015}
 
 \AtBeginSection[]
 {
@@ -124,21 +124,70 @@
   	\end{itemize}
   \end{frame}
   
-\section{Architecture of rsdmx}
-  
-  %rsdmx package%
-  %-------------%
+  %rsdmx package introduction - 1%
+  %------------------------------%
   \begin{frame}
   	\frametitle{rsdmx}
-  	\framesubtitle{A package for reading SDMX data and metadata documents in R}
-  	The rsdmx package:
+  	\framesubtitle{Introduction}
+
   	\begin{itemize}
-		\item provides a set of tools to read SDMX-ML data and metadata documents
-		\item focuses on the SDMX-ML format (main SDMX format)
-		\item provides a single end-user \texttt{readSDMX} function to read any SDMX-ML document, whatever its type
-		\item provides a set of methods to convert SDMX data into tabular data (\texttt{data.frame})
+		\item First initiative to read SDMX in R made available to the R community		
+		\item Set of tools to read SDMX-ML data and metadata documents				
   	\end{itemize}
+	  	
   \end{frame}
+
+  %rsdmx package introduction - 2%
+  %------------------------------%
+  \begin{frame}
+  	\frametitle{rsdmx}
+  	\framesubtitle{Introduction}
+  	
+  	\begin{itemize}
+		\item Generic SDMX abstraction library for R
+		\item Read SDMX documents in a flexible way:
+			\begin{itemize}
+				\item Support for SDMX-ML 1.0, 2.0 and 2.1 standard formats
+				\item Support for \textit{remote} or \textit{local} sources,
+				\item No restriction to the SDMX web-services standard specifications (for \textit{remote} sources)
+			\end{itemize}
+		\item Variety of SDMX documents:
+			\begin{itemize}
+				\item \textbf{Data} (generic, compact, structure-specific, etc.)
+				\item \textbf{Metadata} (Data structure definition - DSD, Codelists, Concepts, etc.)
+			\end{itemize}						
+  	\end{itemize}
+	  	
+  \end{frame}
+  
+  
+  %rsdmx package introduction - 3%
+  %------------------------------%
+  \begin{frame}
+  	\frametitle{rsdmx}
+  	\framesubtitle{Introduction}
+  	
+  	\begin{multicols}{2}
+  	  	
+  	\begin{itemize}
+		\item a single \texttt{readSDMX} function, with a \textbf{large bandwidth of use}:
+			\begin{itemize}
+				\item "raw" approach (read from \textit{url} or \textit{file})
+				\item "helping" approach (read from a list of well-known service providers, with no need to specify the entire request)
+			\end{itemize}
+		\item a set of generic methods to convert SDMX data into tabular data (\texttt{data.frame})
+  	\end{itemize}
+  	
+  	\columnbreak
+	\hspace{2cm}
+	\raggedleft
+	\includegraphics[width=6cm]{img/rsdmx.png}
+
+	\end{multicols}
+	  	
+  \end{frame}
+ 
+\section{Architecture of rsdmx} 
   
   %rsdmx package - S4 model%
   %------------------------%
@@ -197,6 +246,8 @@ Known Subclasses:
 		\begin{itemize}
 			\item \texttt{GenericData}: generic SDMX data format
 			\item \texttt{CompactData}: compacted data format
[…]
