### 2015-09-30 11e61a9
Update quickstart.Rmd
