+++ vignettes/quickstart.Rmd
@@ -56,6 +56,8 @@ This section will introduce you on how to read SDMX *dataset* documents, either 
 
 ### Read _remote_ datasets
 
+#### using the _raw_ approach (specifying the complete request URL)
+
 The following code snipet shows you how to read a dataset from a remote data source, taking as example the [FAO data portal](http://data.fao.org/sdmx/index.html): [http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008](http://data.fao.org/sdmx/repository/data/CROP_PRODUCTION/.156.5312../FAO?startPeriod=2008&endPeriod=2008)
 
 ```{r}
@@ -73,7 +75,16 @@ You can try it out with other datasources, such as:
 
 The online rsdmx documentation also provides a list of data providers, either from international or national institutions.
 
-Now, the service providers above mentioned are known by ``rsdmx`` which let users using ``readSDMX`` with the helper parameters. Let's see how it would look like for querying an OECD datasource:
+#### using the _helper_ approach
+
+Now, the service providers above mentioned are known by ``rsdmx`` which let users using ``readSDMX`` with the helper parameters. The list of service providers can be retrieved doing:
+
+```{r, echo = FALSE}
+providers <- getSDMXServiceProviders();
+sapply(providers, slot, "agencyId")
+```
+
+Then let's see how it would look like for querying an OECD datasource:
 
 ```{r, echo = FALSE}
 sdmx <- readSDMX(id = "OECD", operation = "GetData", key = "MIG", filter = "TOT..", start = 2011, end = 2011)
