### 2015-09-30 9326989
Update quickstart.Rmd
