+++ vignettes/quickstart.Rmd
@@ -84,10 +84,13 @@ providers <- getSDMXServiceProviders();
 sapply(providers, slot, "agencyId")
 ```
 
-Then let's see how it would look like for querying an OECD datasource:
+Note it is also possible to add an SDMX service provider at runtime. For registering a new SDMX service provider by default, please contact me!
+
+Let's see how it would look like for querying an ``OECD`` datasource:
 
 ```{r, echo = FALSE}
-sdmx <- readSDMX(id = "OECD", operation = "GetData", key = "MIG", filter = "TOT..", start = 2011, end = 2011)
+sdmx <- readSDMX(agencyId = "OECD", operation = "GetData", key = "MIG",
+                filter = list("TOT", NULL, NULL), start = 2011, end = 2011)
 df <- as.data.frame(sdmx)
 head(df)
 ```
