### 2015-10-02 8b94d34
Note on role of rsdmx to collate scattered data
