### 2015-10-02 3d4da7f
#66 fix vignette building
