+++ man/rsdmx-package.Rd
@@ -16,7 +16,7 @@ Package: \tab rsdmx\cr
 Type: \tab Package\cr
 Version
 : \tab 0.5-0\cr
-Date: \tab 2015-09-24\cr
+Date: \tab 2015-10-01\cr
 License: \tab GPL(>=2.0)\cr
 LazyLoad: \tab yes\cr
 }
