+++ doc/slides/rsdmx_presentation.tex
@@ -181,7 +181,7 @@
   	\columnbreak
 	\hspace{2cm}
 	\raggedleft
-	\includegraphics[width=6cm]{img/rsdmx.png}
+	\includegraphics[width=6cm]{./img/rsdmx.png}
 
 	\end{multicols}
 	  	
