### 2015-10-05 0810a3b
remove FAO source from vignette & readme
