### 2015-10-05 899c2e1
deactivate some tests for travis and cran
