### 2015-10-05 31cb126
remove most of examples from quickstart, blocking for CRAN because temporary service unavailability...!!
