### 2015-10-06 9019f26
fix code examples for vignette html
