+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rsdmx
-Version: 0.5-0
-Date: 2015-10-01
+Version: 0.5-1
+Date: 2015-10-06
 Title: Tools for Reading SDMX Data and Metadata
 Authors@R: c(
   person("Emmanuel", "Blondel", role = c("aut", "cre"), email = "emmanuel.blondel1@gmail.com"),
