### 2015-10-06 3a5f365
#67 fix unit test
