### 2015-10-06 f742004
#67 #68 enhance REST service interface
