+++ tests/testthat/test_RequestBuilder.R
@@ -11,35 +11,35 @@ test_that("SDMXRequestBuilder",{
   
   request <- SDMXRequestBuilder(
     baseUrl = "http://www.myorg.org",
-    suffix = TRUE,
-    handler = function(baseUrl, agencyId, suffix, operation, key, filter, start, end){
-      return(paste(baseUrl, agencyId, operation, key, filter, start, end, sep="/"))
-    })
+    handler = function(baseUrl, agencyId, resource, flowRef, key, start, end, compliant){
+      return(paste(baseUrl, agencyId, resource, flowRef, key, start, end, sep="/"))
+    },
+    compliant = TRUE)
   
   expect_is(request, "SDMXRequestBuilder")
   expect_equal(request@baseUrl, "http://www.myorg.org")
-  expect_equal(request@suffix, TRUE)
   expect_is(request@handler, "function")
+  expect_equal(request@compliant, TRUE)
   
-  webRequest <- request@handler(baseUrl = "http://www.myorg.org", agencyId = "MYORG", suffix = TRUE, operation = "data", key = "KEY",
-                                filter = "FILTER", start = 2000, end = 2010)
-  expect_equal(webRequest, "http://www.myorg.org/MYORG/data/KEY/FILTER/2000/2010")
+  webRequest <- request@handler(baseUrl = "http://www.myorg.org", agencyId = "MYORG", resource = "data", flowRef = "FLOW",
+                                key = "KEY", start = 2000, end = 2010, compliant = TRUE)
+  expect_equal(webRequest, "http://www.myorg.org/MYORG/data/FLOW/KEY/2000/2010")
   
 })
 
 
 test_that("SDMXRESTRequestBuilder",{
   
-  request <- SDMXRESTRequestBuilder(baseUrl = "http://www.myorg.org", suffix = TRUE)
+  request <- SDMXRESTRequestBuilder(baseUrl = "http://www.myorg.org", compliant = TRUE)
   
   expect_is(request, "SDMXRESTRequestBuilder")
   expect_equal(request@baseUrl, "http://www.myorg.org")
-  expect_equal(request@suffix, TRUE)
   expect_is(request@handler, "function")
+  expect_equal(request@compliant, TRUE)
   
-  webRequest <- request@handler(baseUrl = "http://www.myorg.org", agencyId = "MYORG", suffix = TRUE, operation = "data", key = "KEY",
-                                filter = "FILTER", start = 2000, end = 2010)
-  expect_equal(webRequest, "http://www.myorg.org/data/KEY/FILTER/MYORG?startPeriod=2000&endPeriod=2010")
+  webRequest <- request@handler(baseUrl = "http://www.myorg.org", agencyId = "MYORG", resource = "data", flowRef = "FLOW",
+                                key = "KEY", start = 2000, end = 2010, compliant = TRUE)
+  expect_equal(webRequest, "http://www.myorg.org/data/FLOW/KEY/?startPeriod=2000&endPeriod=2010")
   
 })
 
