+++ tests/testthat/test_ServiceProvider.R
@@ -12,14 +12,14 @@ test_that("SDMXServiceProvider - constructor",{
   requestBuilder <- SDMXRequestBuilder(
     baseUrl = "http://www.myorg.org",
     suffix = TRUE,
-    handler = function(baseUrl, agencyId, suffix, operation, key, filter, start, end){
-      return(paste(baseUrl, agencyId, operation, key, filter, start, end, sep="/"))
+    handler = function(baseUrl, agencyId, resource, flowRef, key, start, end, compliant){
+      return(paste(baseUrl, agencyId, resource, flowRef, key, start, end, sep="/"))
     })
   
   
   provider <- SDMXServiceProvider(
     agencyId = "MYORG", name = "My Organization",
-    requestBuilder
+    builder = requestBuilder
   )
   
   expect_is(provider, "SDMXServiceProvider")
@@ -34,7 +34,7 @@ test_that("SDMXServiceProvider - methods",{
     
   providers <- getSDMXServiceProviders()
   expect_is(providers, "list")
-  expect_equal(length(providers), 5L)
+  expect_equal(length(providers), 6L)
   expect_equal(sapply(providers, function(x){slot(x,"agencyId")}),
                c("ECB", "ESTAT", "OECD", "FAO", "ILO", "UIS"))
   
@@ -48,12 +48,12 @@ test_that("SDMXServiceProvider - methods",{
   
   provider <- SDMXServiceProvider(
     agencyId = "MYORG", name = "My Organization",
-    requestBuilder
+    builder = requestBuilder
   )
   
   addSDMXServiceProvider(provider)
   providers <- getSDMXServiceProviders()
-  expect_equal(length(providers), 6L)
+  expect_equal(length(providers), 7L)
   expect_equal(sapply(providers, function(x){slot(x,"agencyId")}),
                c("ECB", "ESTAT", "OECD", "FAO", "ILO", "UIS", "MYORG"))
   
