### 2015-10-06 1a40f57
#70 additional providers properties (scale, country), #71 wrapping class + as.data.frame method for listing available providers
