+++ R/Class-SDMXServiceProvider.R
@@ -6,11 +6,15 @@ setClass("SDMXServiceProvider",
          representation(
            agencyId = "character",
            name = "character",
+           scale = "character",
+           country = "character",
            builder = "SDMXRequestBuilder"
          ),
          prototype = list(
            agencyId = "MYORG",
            name = "My Organization",
+           scale = "international",
+           country = as.character(NA),
            footer = new("SDMXRequestBuilder")
          ),
          validity = function(object){
