### 2015-10-06 f121052
#72 support for REST 'datastructure' resource
