### 2015-10-06 6dfbc41
ignore building vignettes for travis-ci
