### 2015-10-07 e68dad6
exclude setSDMXServiceProviders() from cov stats
