### 2015-10-07 a78bbbc
ignore .onLoad from cov stats
