+++ R/profile.R
@@ -1,4 +1,4 @@
-.onLoad <- function (libname, pkgname) {
+.onLoad <- function (libname, pkgname) { # nocov start
   assign(".rsdmx.options", new.env(), envir= asNamespace(pkgname))
   setSDMXServiceProviders()
-}
+} # nocov end
