### 2015-10-07 b176034
#74 add INSEE (France) web-service connector
