+++ R/SDMXServiceProviders-methods.R
@@ -14,7 +14,8 @@ as.data.frame.SDMXServiceProviders <- function(x, ...){
                       slot(provider, "scale"), slot(provider, "country"))
                   })),
             stringsAsFactors = FALSE)
-  colnames(out) <- c("agencyId", "name", "scale", "country")  
+  colnames(out) <- c("agencyId", "name", "scale", "country")
+  Encoding(out$name) <- "UTF-8"
   return(out)
 }
 
