### 2015-10-07 6f9929a
fix unit tests
