+++ R/Class-SDMXRESTRequestBuilder.R
@@ -4,8 +4,12 @@
 #SDMX REST Request builder class (still experimental)
 setClass("SDMXRESTRequestBuilder",
          contains = "SDMXRequestBuilder",
-         representation(),
-         prototype = list(),
+         representation(
+          forceAgencyId = "logical"  
+         ),
+         prototype = list(
+          forceAgencyId = FALSE
+         ),
          validity = function(object){
            #eventual validation rules
            return(TRUE);
