+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rsdmx
 Version: 0.5-1
-Date: 2015-10-06
+Date: 2015-10-08
 Title: Tools for Reading SDMX Data and Metadata
 Authors@R: c(
   person("Emmanuel", "Blondel", role = c("aut", "cre"), email = "emmanuel.blondel1@gmail.com"),
@@ -17,6 +17,7 @@ VignetteBuilder: knitr
 URL: https://github.com/opensdmx/rsdmx, http://www.sdmx.org
 BugReports: https://github.com/opensdmx/rsdmx/issues
 LazyLoad: yes
+Encoding: UTF-8
 Collate:
   Class-SDMXSchema.R Class-SDMXType.R Class-SDMXStructureType.R
   Class-SDMXHeader.R Class-SDMXFooterMessage.R Class-SDMXFooter.R
