+++ R/SDMXRESTRequestBuilder-methods.R
@@ -2,7 +2,8 @@
 #========================
 
 #constructor
-SDMXRESTRequestBuilder <- function(regUrl, repoUrl, compliant, forceAgencyId = FALSE){
+SDMXRESTRequestBuilder <- function(regUrl, repoUrl, compliant,
+                                   skipAgencyId = FALSE, forceAgencyId = FALSE){
   
   #function to handle request
   serviceRequestHandler <- function(regUrl, repoUrl, agencyId, resource, resourceId, version = NULL,
@@ -35,19 +36,23 @@ SDMXRESTRequestBuilder <- function(regUrl, repoUrl, compliant, forceAgencyId = F
         
         #base data request
         req <- paste(obj$repoUrl, obj$resource, obj$flowRef, obj$key, sep = "/")
-        if(forceAgencyId) req <- paste(req, obj$agencyId, sep = "/")
+        if(skipAgencyId){
+          req <- paste0(req, "/")
+        }else{
+          req <- paste(req, ifelse(forceAgencyId, obj$agencyId, "all"), sep = "/")
+        }
         
         #DataQuery
         #-> temporal extent (if any)
         addParams = FALSE
         if(!is.null(obj$start)){
-          req <- paste0(req, ifelse(forceAgencyId, "", "/"), "?")
+          req <- paste0(req, "?")
           addParams = TRUE
           req <- paste0(req, "startPeriod=", start)
         }
         if(!is.null(obj$end)){
           if(!addParams){
-            req <- paste0(req, ifelse(forceAgencyId, "", "/"), "?")
+            req <- paste0(req, "?")
           }else{
             req <- paste0(req, "&")
           }
@@ -86,7 +91,9 @@ SDMXRESTRequestBuilder <- function(regUrl, repoUrl, compliant, forceAgencyId = F
       regUrl = regUrl,
       repoUrl = repoUrl,
       handler = serviceRequestHandler,
-      compliant = compliant)
+      compliant = compliant,
+      skipAgencyId = skipAgencyId,
+      forceAgencyId = forceAgencyId)
 }
 
 
