+++ tests/testthat/test_Main_Helpers.R
@@ -13,6 +13,18 @@ context("SDMXHelpers")
 
 #ECB
 #---
+
+#-> datastructure
+test_that("ECB - datastructure",{
+  testthat::skip_on_travis()
+  testthat::skip_on_cran()
+  sdmx <- readSDMX(agencyId = "ECB", resource = "datastructure", resourceId = "ECB_DD1")
+  if(!is.null(sdmx)){
+    expect_is(sdmx, "SDMXDataStructureDefinition")
+  }
+})
+
+#-> data
 test_that("ECB - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
@@ -24,17 +36,20 @@ test_that("ECB - data",{
   }
 })
 
-test_that("ECB - datastructure",{
+#ESTAT (EUROSTAT)
+#---------------
+
+#-> datastructure
+test_that("ESTAT - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ECB", resource = "datastructure", resourceId = "ECB_DD1")
+  sdmx <- readSDMX(agencyId = "ESTAT", resource = "datastructure", resourceId = "DSD_nama_gdp_c")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataStructureDefinition")
   }
 })
 
-#ESTAT (EUROSTAT)
-#---------------
+#-> data
 test_that("ESTAT - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
@@ -46,17 +61,20 @@ test_that("ESTAT - data",{
   }
 })
 
-test_that("ESTAT - datastructure",{
+#OECD
+#----
+
+#-> datastructure
+test_that("OECD - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ESTAT", resource = "datastructure", resourceId = "DSD_nama_gdp_c")
+  sdmx <- readSDMX(agencyId = "OECD", resource = "datastructure", resourceId = "TABLE1")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataStructureDefinition")
   }
 })
 
-#OECD
-#----
+#-> data
 test_that("OECD - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
@@ -67,17 +85,20 @@ test_that("OECD - data",{
   }
 })
 
-test_that("OECD - datastructure",{
+#FAO (UN-FAO)
+#------------
+
+#-> datastructure
+test_that("FAO - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "OECD", resource = "datastructure", resourceId = "TABLE1")
+  sdmx <- readSDMX(agencyId = "FAO", resource = "datastructure", resourceId = "FAOSTAT")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataStructureDefinition")
   }
 })
 
-#FAO (UN-FAO)
-#------------
+#-> data
 test_that("FAO - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
@@ -89,17 +110,20 @@ test_that("FAO - data",{
   }
 })
 
-test_that("FAO - datastructure",{
+#ILO (UN-ILO)
+#------------
+
+#-> datastructure
+test_that("ILO - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "FAO", resource = "datastructure", resourceId = "FAOSTAT")
[…]
