+++ R/SDMXStructureType-methods.R
@@ -1,13 +1,13 @@
 # E.Blondel - 2014/08/03
 #=======================
 
-SDMXStructureType <- function(xmlObj){
+SDMXStructureType <- function(xmlObj, resource){
 	new("SDMXStructureType",
       SDMXType(xmlObj),
-      subtype = type.SDMXStructureType(xmlObj));
+      subtype = type.SDMXStructureType(xmlObj, resource));
 }
 
-type.SDMXStructureType <- function(xmlObj){
+type.SDMXStructureType <- function(xmlObj, resource){
   
   sdmxVersion <- version.SDMXSchema(xmlObj)
   VERSION.21 <- sdmxVersion == "2.1"
@@ -17,6 +17,7 @@ type.SDMXStructureType <- function(xmlObj){
   strNs <- findNamespace(namespaces, "structure")
   
   if(VERSION.21){
+    flowXML <- getNodeSet(xmlObj, "//ns:DataFlows", namespaces = strNs)
     dsXML <- getNodeSet(xmlObj, "//ns:DataStructures", namespaces = strNs)
     ccXML <- getNodeSet(xmlObj, "//ns:Concepts", namespaces = strNs)
     clXML <- getNodeSet(xmlObj, "//ns:Codelists", namespaces = strNs)
@@ -30,6 +31,7 @@ type.SDMXStructureType <- function(xmlObj){
       return(strType)
     }
   }else{
+    #TODO flowXML
     dsXML <- getNodeSet(xmlObj, "//ns:KeyFamilies", namespaces = messageNs)
     ccXML <- getNodeSet(xmlObj, "//ns:Concepts", namespaces = messageNs)
     clXML <- getNodeSet(xmlObj, "//ns:CodeLists", namespaces = messageNs)
@@ -40,7 +42,13 @@ type.SDMXStructureType <- function(xmlObj){
       #others
       if(length(ccXML)>0) return("ConceptsType")
       if(length(clXML)>0) return("CodelistsType")
-      if(length(dsXML)>0) return("DataStructuresType")
+      if(length(dsXML)>0){
+        strType <- switch(resource,
+                    "dataflow" = "DataflowsType",
+                    "datastructure" = "DataStructuresType",
+                    NULL)
+        return(strType)
+      }
     }
   }
   return(NULL)
