### 2015-10-19 01a0bb3
#9 start oxygenizing Rdoc
