+++ R/SDMXCompactData-methods.R
@@ -12,8 +12,8 @@ as.data.frame.SDMXAllCompactData <- function(x, nsExpr, ...) {
   xmlObj <- x@xmlObj;
   dataset <- NULL
   
-  schema <- getSDMXSchema(x)
-  sdmxVersion <- getVersion(schema)
+  schema <- slot(x,"schema")
+  sdmxVersion <- slot(schema,"version")
   VERSION.21 <- sdmxVersion == "2.1"
   
   #namespace
