+++ R/Class-SDMX.R
@@ -1,7 +1,31 @@
-# E.Blondel - 2013/06/09
-#=======================
+#' @name SDMX
+#' @rdname SDMX
+#' @docType class
+#' @aliases SDMX-class
+#' 
+#' @title Class "SDMX"
+#' @description An abstract class from which SDMX classes are derived
+#'
+#' @slot xmlObj Object of class "XMLInternalDocument" derived from XML package  
+#' @slot schema Object of class "SDMXSchema", handles the version of SDMX-ML format
+#' @slot header Object of class "SDMXHeader", handles the SDMX-ML document header
+#' @slot footer Object of class "SDMXFooter", handles the SDMX-ML document footer
+#'  
+#' @section Warning:
+#' This class is not useful in itself, but all SDMX classes in this package derive
+#' from it.                  
+#'  
+#' @note
+#'  Currently, the approach drafted in \link{rsdmx} package was to rely on XML
+#'  package, read the xml object and store it as part of the SDMX R object. Another
+#'  approach being investigated is to use XML handlers throughthe Simple API for XML
+#'  (SAX) that could avoid to load the full XML tree in the SDMX R object (xmlObj). 
+#'  Indeed, SDMX data could be huge and causes issues of memory if the complete XML
+#'  tree is loaded in the R user session.
+#'  
+#'  @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#'                  
 
-#SDMX base abstract class
 setClass("SDMX",
 	representation(
 		xmlObj = "XMLInternalDocument",
