+++ R/SDMX-methods.R
@@ -1,6 +1,14 @@
-# constructor
-# E.Blondel - 2013/06/09
-#=======================
+#' @name SDMX
+#' @rdname SDMX
+#' @aliases SDMX,SDMX-method
+#' 
+#' @usage
+#' SDMX(xmlObj)
+#' 
+#' @param xmlObj object of class "XMLInternalDocument derived from XML package
+#' @return an object of class "SDMX"
+#' 
+#' @seealso \link{readSDMX}
 
 SDMX <- function(xmlObj){
 	schema <- SDMXSchema(xmlObj);
@@ -13,49 +21,7 @@ SDMX <- function(xmlObj){
       footer = footer); 
 }
 
-
-#generics
-if (!isGeneric("as.XML"))
-	setGeneric("as.XML", function(obj) standardGeneric("as.XML"));
-	
-if (!isGeneric("getSDMXSchema"))
-	setGeneric("getSDMXSchema", function(obj) standardGeneric("getSDMXSchema"));
-	
-if (!isGeneric("getSDMXHeader"))
-	setGeneric("getSDMXHeader", function(obj) standardGeneric("getSDMXHeader"));
-	
-if (!isGeneric("getSDMXType"))
-	setGeneric("getSDMXType", function(obj) standardGeneric("getSDMXType"));
-
-if (!isGeneric("getNamespaces"))
-  setGeneric("getNamespaces", function(obj) standardGeneric("getNamespaces"));
-
-if (!isGeneric("getSDMXFooter"))
-  setGeneric("getSDMXFooter", function(obj) standardGeneric("getSDMXFooter"));
-
-#methods
-setMethod(f = "as.XML", signature = "SDMX", function(obj){
-									return(obj@xmlObj);
-							}
-)
-setMethod(f = "getSDMXSchema", signature = "SDMX", function(obj){
-									return(obj@schema);
-							}
-)
-setMethod(f = "getSDMXHeader", signature = "SDMX", function(obj){
-									return(obj@header);
-							}
-)
-setMethod(f = "getSDMXType", signature = "SDMX", function(obj){
-									return(SDMXType(obj@xmlObj));
-							}
-)
-setMethod(f = "getSDMXFooter", signature = "SDMX", function(obj){
-                return(SDMXFooter(obj@xmlObj));
-              }
-)
-
-
+#functions
 namespaces.SDMX <- function(xmlObj){
   nsFromXML <- xmlNamespaceDefinitions(xmlObj, recursive = TRUE, simplify = FALSE)
   nsDefs.df <- do.call("rbind",
@@ -75,29 +41,107 @@ namespaces.SDMX <- function(xmlObj){
   return(nsDefs.df)
 }
 
+#' @name getNamespaces
+#' @docType methods
+#' @aliases getNamespaces,SDMX-method
+#' @title getNamespaces
+#' @description Access the namespaces of the SDMX-ML object
+#' @usage getNamespaces(obj)
+#' 
+#' @param obj An object deriving from class "SDMX"
+#' @return an object of class \code{data.frame} giving the id and uri for each 
+#'         of the namespaces handled in the SDMX-ML document.
+#'
+#' @seealso \link{SDMX-class}
+#'
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+
+if (!isGeneric("getNamespaces"))
+  setGeneric("getNamespaces", function(obj) standardGeneric("getNamespaces"));
+
+#' @describeIn getNamespaces
 setMethod(f = "getNamespaces", signature = "SDMX", function(obj){
             return(namespaces.SDMX(obj@xmlObj));
-          }
-)
+          })
 
 #others non-S4 methods
 #====================
 
-#findNamespace
+#' @name findNamespace
+#' @aliases findNamespace
+#' @title findNamespace
[…]
