### 2015-10-26 bdc7c7d
fix with removing getType generic
