+++ R/Class-SDMXSchema.R
@@ -6,8 +6,9 @@
 #' 
 #' @slot version Object of class "character" giving the SDMX-ML schema version 
 #' 
-#' @section Warning: this class is not useful in itself, but all SDMX non-abstract
-#'          classes will encapsulate it as slot, when parsing an SDMX-ML document.
+#' @section Warning:
+#' this class is not useful in itself, but all SDMX non-abstract classes will 
+#' encapsulate it as slot, when parsing an SDMX-ML document.
 #'          
 #' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
 #' 
@@ -17,7 +18,7 @@ setClass("SDMXSchema",
 		),
 		prototype = list(version = "2.0"),
 		validity = function(object){
-			VERSION <- getVersion(object);
+			VERSION <- object@version;
 			valid <- switch(VERSION,
                 "1.0" = TRUE,
                 "2.0" = TRUE,
