+++ man/SDMSchema.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/SDMXSchema-methods.R
-\name{SDMXSchema}
-\alias{SDMXSchema}
-\alias{SDMXSchema,SDMXSchema-method}
-\usage{
-SDMXSchema(xmlObj)
-}
-\arguments{
-\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
-}
-\value{
-an object of class "SDMXSchema"
-}
-\seealso{
-\link{readSDMX}
-}
-
