### 2015-10-26 220939c
#9 oxygenize SDMXType Rdoc
