### 2015-10-26 71a8a12
delete unused Rdoc
