### 2015-10-26 28976fd
#9 more oxygen + fix unit tests
