+++ man/rsdmx-package.Rd
@@ -1,31 +1,0 @@
-\name{rsdmx}
-\Rdversion{1.1}
-\alias{rsdmx-package}
-\alias{rsdmx}
-
-
-\docType{package}
-
-\title{Tools for Reading SDMX Data and Metadata}
-
-\description{RSDMX is a package to parse/read SDMX documents in R. It provides a set of classes and methods to read data and metadata documents exchanged through the Statistical Data and Metadata Exchange (SDMX) framework. The package currently focuses on the SDMX XML standard format (SDMX-ML).}
-
-\details{
-\tabular{ll}{
-Package: \tab rsdmx\cr
-Type: \tab Package\cr
-Version
-: \tab 0.5-1\cr
-Date: \tab 2015-10-08\cr
-License: \tab GPL(>=2.0)\cr
-LazyLoad: \tab yes\cr
-}
-
-}
-
-\author{Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}}
-
-
-\keyword{SDMX}
-\keyword{statistics}
-\keyword{timeseries}

[…]
