+++ R/Class-SDMXDimension.R
@@ -1,7 +1,36 @@
-# E.Blondel - 2014/08/20
-#=======================
-
-#SDMX Components class
+#' @name SDMXDimension
+#' @docType class
+#' @aliases SDMXDimension-class
+#' 
+#' @title Class "SDMXDimension"
+#' @description A basic class to handle a SDMX Dimension
+#'
+#' @slot conceptRef Object of class "character" giving the dimension conceptRef (required)
+#' @slot conceptVersion Object of class "character" giving the dimension concept version
+#' @slot conceptAgency Object of class "character" giving the dimension concept agency
+#' @slot conceptSchemeRef Object of class "character" giving the dimension conceptScheme ref
+#' @slot conceptSchemeAgency Object of class "character" giving the dimension conceptScheme agency
+#' @slot codelist Object of class "character" giving the codelist ref name
+#' @slot codelistVersion Object of class "character" giving the codelist ref version
+#' @slot codelistAgency Object of class "character" giving the codelist ref agency
+#' @slot isMeasureDimension Object of class "logical" indicating if the dimension is a measure dimension. Default value is FALSE
+#' @slot isFrequencyDimension Object of class "logical" indicating if the dimension is a frequency dimension. Default value is FALSE
+#' @slot isEntityDimension Object of class "logical" indicating if the dimension is an entity dimension. Default value is FALSE
+#' @slot isCountDimension Object of class "logical" indicating if the dimension is a count dimension. Default value is FALSE
+#' @slot isNonObservationTimeDimension Object of class "logical" indicating if the dimension is a non-observation dimension. Default value is FALSE
+#' @slot isIdentityDimension Object of class "logical" indicating if the dimension is an identity dimension. Default value is FALSE
+#' @slot crossSectionalAttachDataset Object of class "logical"
+#' @slot crossSectionalAttachGroup Object of class "logical"
+#' @slot crossSectionalAttachSection Object of class "logical"
+#' @slot crossSectionalAttachObservation Object of class "logical"         
+#'
+#' @section Warning:
+#' This class is not useful in itself, but all SDMX non-abstract classes will 
+#' encapsulate it as slot, when parsing an SDMX-ML document (Concepts, or 
+#' DataStructureDefinition)
+#'    
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#'
 setClass("SDMXDimension",
          representation(
           #attributes
