+++ R/Class-SDMXRESTRequestBuilder.R
@@ -1,7 +1,20 @@
-# E.Blondel - 2015/09/22
-#=======================
-
-#SDMX REST Request builder class (still experimental)
+#' @name SDMXRESTRequestBuilder
+#' @docType class
+#' @aliases SDMXRESTRequestBuilder-class
+#' 
+#' @title Class "SDMXRESTRequestBuilder"
+#' @description A experimental class to handle a SDMX 2.1 service request builder
+#' 
+#' @slot regUrl an object of class "character" giving the base Url of the SDMX service registry
+#' @slot repoUrl an object of class "character" giving the base Url of the SDMX service repository
+#' @slot compliant an object of class "logical" indicating if the web-service is compliant with the SDMX REST web-service specifications
+#'
+#' @section Warning:
+#' This class is not useful in itself, but all SDMX non-abstract classes will 
+#' encapsulate it as slot, when parsing an SDMX-ML document.
+#' 
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#' 
 setClass("SDMXRESTRequestBuilder",
          contains = "SDMXRequestBuilder",
          representation(),
