### 2015-10-27 a7f9b01
description - web-service interface facility
