### 2015-12-08 4e0d319
fix in case of Sender is null
although a sender should not be null according to SDMX specs, WorldBank let user filling (or not) the sender/receiver. If user doesn't specify anything rsdmx breaks. Just added as convenience (but misalignement with SDMX std)
