### 2015-12-08 1e37699
convenience in case receiver/sender null (mainly for WorldBank data downloaders)
