### 2016-01-18 02435dd
#80 fix minor bug in case of multiple authority namespaces
