### 2016-01-19 035186d
#81 tentative support for CrossSectionalData
