+++ R/rsdmx.R
@@ -17,7 +17,7 @@
 #'    Type: \tab Package\cr
 #'    Version
 #'    : \tab 0.5-1\cr
-#'    Date: \tab 2015-10-08\cr
+#'    Date: \tab 2016-01-19\cr
 #'    License: \tab GPL(>=2.0)\cr
 #'    LazyLoad: \tab yes\cr
 #'  }
