+++ man/SDMXCrossSectionalData.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/Class-SDMXCrossSectionalData.R, R/SDMXCrossSectionalData-methods.R
+\docType{class}
+\name{SDMXCrossSectionalData}
+\alias{SDMXCrossSectionalData}
+\alias{SDMXCrossSectionalData,SDMXCrossSectionalData-method}
+\alias{SDMXCrossSectionalData-class}
+\title{Class "SDMXCrossSectionalData"}
+\usage{
+SDMXCrossSectionalData(xmlObj)
+}
+\arguments{
+\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
+}
+\value{
+an object of class "SDMXCrossSectionalData"
+}
+\description{
+A basic class to handle a SDMX-ML cross sectional data set
+}
+\author{
+Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
+}
+\seealso{
+\link{readSDMX}
+}
+

[…]
