### 2016-01-19 7a04ba0
fix regression in compactdata parser
