### 2016-01-20 b674755
no need of authorityNs logic in GenericDat parser
