+++ R/SDMXGenericData-methods.R
@@ -28,21 +28,6 @@ as.data.frame.SDMXGenericData <- function(x, ...){
   #namespace
   nsDefs.df <- getNamespaces(x)
   ns <- findNamespace(nsDefs.df, "generic")
-  if(length(ns) == 0){
-    #in case no ns found, try to find specific namespace
-    ns.df <- nsDefs.df[
-      regexpr("http://www.sdmx.org", nsDefs.df$uri,
-              "match.length", ignore.case = TRUE) == -1
-      & regexpr("http://www.w3.org", nsDefs.df$uri,
-                "match.length", ignore.case = TRUE) == -1,]
-    ns <- ns.df$uri
-    if(length(ns) > 1){
-      warning("More than one target dataset namespace found!")
-      ns <- ns[1L]
-    }
-    hasAuthorityNS <- TRUE
-    authorityId <- nsDefs.df[nsDefs.df$uri == ns,]$id
-  }
   
   #series
   seriesXML <- getNodeSet(xmlObj, "//ns:Series", namespaces = ns)
