### 2016-02-07 55c5876
minor fixes
