### 2016-02-07 4e3a8ed
#82 add KNOEMA test service provider
