+++ R/rsdmx.R
@@ -16,8 +16,8 @@
 #'    Package: \tab rsdmx\cr
 #'    Type: \tab Package\cr
 #'    Version
-#'    : \tab 0.5-1\cr
-#'    Date: \tab 2016-01-19\cr
+#'    : \tab 0.5-2\cr
+#'    Date: \tab 2016-02-07\cr
 #'    License: \tab GPL(>=2.0)\cr
 #'    LazyLoad: \tab yes\cr
 #'  }
