### 2016-02-07 be40de6
#82 fix unit tests
