### 2016-02-08 5499877
simplify service provider tests
