### 2016-02-10 7495cf3
#67 supports now 'dataflow' resource
