+++ R/Class-SDMXMessageGroup.R
@@ -12,7 +12,7 @@
 #' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
 #' 
 setClass("SDMXMessageGroup",
-         contains = "SDMX",
+         contains = "SDMXData",
          representation(),
          prototype = list(),
          validity = function(object){
