+++ R/Class-SDMXCompactData.R
@@ -8,7 +8,7 @@
 #' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
 #' 
 setClass("SDMXCompactData",
-         contains = "SDMX",
+         contains = "SDMXData",
          representation(),
          prototype = list(),
          validity = function(object){
