### 2016-02-10 c19e72e
#83 fix Rdoc
