### 2016-02-10 ad9f87b
#85 support for SDMXComponents as.data.frame method
