+++ NAMESPACE
@@ -101,5 +101,6 @@ S3method(as.data.frame, SDMXMessageGroup)
 S3method(as.data.frame, SDMXConcepts)
 S3method(as.data.frame, SDMXCodelists)
 S3method(as.data.frame, SDMXDataStructures)
+S3method(as.data.frame, SDMXComponents)
 S3method(as.data.frame, SDMXDataFlows)
 S3method(as.data.frame, SDMXServiceProviders)
