### 2016-02-10 09b2c4c
fix generated Rdoc
