+++ R/SDMXData-methods.R
@@ -50,6 +50,39 @@ dsdRef.SDMXData <- function(xmlObj){
   return(dsdRef)
 }
 
+#ENRICH DATA WITH LABELS
+#=======================
+addLabels.SDMXData <- function(data, dsd){
+  
+  ds <- slot(slot(dsd,"datastructures"), "datastructures")[[1]]
+  components <- slot(ds, "Components")
+  components <- as.data.frame(components)
+  
+  #function to enrich a column with its labels
+  enrichColumnWithLabels <- function(column, dsd, components){
+    
+    datac <- as.data.frame(data[,column], stringsAsFactors = FALSE)
+    colnames(datac) <- column
+    clName <- components[components$conceptRef == column, "codelist"]
+    if(length(clName) != 0 && !is.na(clName) && !is.null(clName)){
+      cl <- as.data.frame(slot(sdmx.dsd, "codelists"), codelistId = clName)
+      datac = merge(x = datac, y = cl, by.x = column, by.y = "id",
+                    all.x = TRUE, all.y = FALSE)
+      datac <- datac[,((regexpr("label", colnames(datac)) != -1) + 
+                         (colnames(datac) == column) == 1)]
+      colnames(datac)[regexpr("label",colnames(datac)) != -1] <- paste0(column,
+      "_",colnames(datac)[regexpr("label",colnames(datac)) != -1])
+    }
+    
+    return(datac)
+    
+  }
+  
+  fulldata <- do.call("cbind" ,lapply(columns, enrichColumnWithLabels,
+                                      dsd, components))
+  return(fulldata)
+}
+
 
 #' @name setDSD
 #' @docType methods
