### 2016-02-10 e3eea1b
#84 fix minor bugs
