+++ R/SDMXCompactData-methods.R
@@ -124,7 +124,6 @@ as.data.frame.SDMXAllCompactData <- function(x, nsExpr, labels = FALSE, ...) {
   
   #enrich with labels
   if(labels){
-    message("we are going to enrich the data.frame")
     dsd <- slot(x, "dsd")
     if(!is.null(dsd)) dataset <- addLabels.SDMXData(dataset, dsd)
   }
