### 2016-02-10 89466da
#86 fix generic data parser
