### 2016-02-10 ef1f251
#87 add support for ENEGI (+improvements)
