+++ R/SDMXDotStatRequestBuilder-methods.R
@@ -1,0 +1,117 @@
+#' @name SDMXDotStatRequestBuilder
+#' @rdname SDMXDotStatRequestBuilder
+#' @aliases SDMXDotStatRequestBuilder,SDMXDotStatRequestBuilder-method
+#' 
+#' @usage
+#'  SDMXDotStatRequestBuilder(regUrl, repoUrl, unsupportedResources,
+#'                            skipAgencyId, forceAgencyId)
+#'
+#' @param regUrl an object of class "character" giving the base Url of the SDMX 
+#'        service registry
+#' @param repoUrl an object of class "character" giving the base Url of the SDMX 
+#'        service repository
+#' @param unsupportedResources an object of class "list" giving eventual unsupported 
+#'        REST resources. Default is an empty list object
+#' @param skipAgencyId an object of class "logical" indicating that agencyId 
+#'        should be skipped. Used to control lack of strong SDMX REST compliance 
+#'        from data providers. For now, it applies only for the "data" resource.
+#' @param forceAgencyId an object of class "logical" indicating if the agencyId 
+#'        as to be added at the end of the request. Default value is \code{FALSE}. 
+#'        For some providers, the \code{all} value for the \code{agencyId} is not 
+#'        allowed, in this case, the \code{agencyId} of the data provider has to 
+#'        be forced in the web-request 
+#'                
+#' @examples
+#'   #how to create a SDMXDotStatRequestBuilder
+#'   requestBuilder <- SDMXDotStatRequestBuilder(
+#'     regUrl = "http://www.myorg/registry",
+#'     repoUrl = "http://www.myorg/repository")
+#'
+SDMXDotStatRequestBuilder <- function(regUrl, repoUrl,
+                                   unsupportedResources = list(),
+                                   skipAgencyId = FALSE, forceAgencyId = FALSE){
+  
+  #function to handle request
+  serviceRequestHandler <- function(regUrl, repoUrl, agencyId, resource, resourceId, version = NULL,
+                                    flowRef, key = NULL, start = NULL, end = NULL,
+                                    compliant){    
+    
+    if(is.null(resource))
+      stop("Missing SDMX service resource")
+    
+    if(resource %in% unsupportedResources)
+      stop("Unsupported SDMX service resource for this provider")
+    
+    #wrap argument values
+    obj <- list(regUrl = regUrl, repoUrl = repoUrl,
+                agencyId = agencyId, resource = resource, resourceId = resourceId, version = version,
+                flowRef = flowRef, key = key, start = start, end = end)
+    
+    #REST resource handler
+    resourceHandler <- switch(resource,
+                              
+      #'dataflow' resource (path="GetKeyFamily/{resourceID}")
+      #------------------------------------------------------
+      "dataflow" = function(obj){
+        if(is.null(obj$resourceId)) obj$resourceId = "all"
+        if(is.null(obj$version)) obj$version = "latest"
+        req <- sprintf("%s/GetKeyFamily/%s",obj$regUrl, obj$resourceId)
+        return(req)
+      },
+      
+      #'datastructure' resource (path="GetDataStructure/{resourceID}/{agencyID}")
+      #--------------------------------------------------------------------------
+      "datastructure" = function(obj){
+        if(is.null(obj$resourceId)) obj$resourceId = "all"
+        if(is.null(obj$version)) obj$version = "latest"
+        req <- sprintf("%s/GetDataStructure/%s",obj$regUrl, obj$resourceId)
+        if(forceAgencyId) req <- paste(req, obj$agencyId, sep = "/")
+        return(req)
+      },
+      
[…]
