+++ man/getRegistryInterfaceResult.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/SDMX-methods.R
+\name{getRegistryInterfaceResult}
+\alias{getRegistryInterfaceResult}
+\title{getRegistryInterfaceResult}
+\usage{
+getRegistryInterfaceResult(xmlObj)
+}
+\arguments{
+\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
+}
+\value{
+an object of class "XMLInternalDocument derived from XML package
+}
+\description{
+function used to extract the SDMX-ML message from a registry
+             interface query
+}
+\section{Warning}{
+
+\code{getRegistryInterfaceResult} is a function used internally by \link{readSDMX}
+}
+\author{
+Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
+}
+\seealso{
+\link{SDMX-class} \link{readSDMX}
+}
+
