### 2016-02-11 c0b2c62
#89 add builder/compliant properties to list of SDMX providers (data.frame)
