+++ R/SDMXServiceProviders-methods.R
@@ -17,11 +17,16 @@ as.data.frame.SDMXServiceProviders <- function(x, ...){
   out <- as.data.frame(do.call("rbind",
            lapply(slot(x, "providers"),
                   function(provider){
+                    
+                    builder <- slot(provider, "builder")
+                    
                     c(slot(provider,"agencyId"), slot(provider, "name"),
-                      slot(provider, "scale"), slot(provider, "country"))
+                      slot(provider, "scale"), slot(provider, "country"),
+                      class(builder), slot(builder, "compliant"))
                   })),
             stringsAsFactors = FALSE)
-  colnames(out) <- c("agencyId", "name", "scale", "country")
+  colnames(out) <- c("agencyId", "name", "scale", "country",
+                     "builder", "compliant")
   return(out)
 }
 
