+++ man/SDMXREST21RequestBuilder.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/Class-SDMXRESTRequestBuilder.R, R/SDMXRESTRequestBuilder-methods.R
+% Please edit documentation in R/Class-SDMXREST21RequestBuilder.R, R/SDMXREST21RequestBuilder-methods.R
 \docType{class}
-\name{SDMXRESTRequestBuilder}
-\alias{SDMXRESTRequestBuilder}
-\alias{SDMXRESTRequestBuilder,SDMXRESTRequestBuilder-method}
-\alias{SDMXRESTRequestBuilder-class}
-\title{Class "SDMXRESTRequestBuilder"}
+\name{SDMXREST21RequestBuilder}
+\alias{SDMXREST21RequestBuilder}
+\alias{SDMXREST21RequestBuilder,SDMXREST21RequestBuilder-method}
+\alias{SDMXREST21RequestBuilder-class}
+\title{Class "SDMXREST21RequestBuilder"}
 \usage{
-SDMXRESTRequestBuilder(regUrl, repoUrl, compliant, unsupportedResources,
+SDMXREST21RequestBuilder(regUrl, repoUrl, compliant, unsupportedResources,
                         skipAgencyId, forceAgencyId)
 }
 \arguments{
@@ -51,8 +51,8 @@ This class is not useful in itself, but all SDMX non-abstract classes will
 encapsulate it as slot, when parsing an SDMX-ML document.
 }
 \examples{
-#how to create a SDMXRESTRequestBuilder
-  requestBuilder <- SDMXRESTRequestBuilder(
+#how to create a SDMXREST21RequestBuilder
+  requestBuilder <- SDMXREST21RequestBuilder(
     regUrl = "http://www.myorg/registry",
     repoUrl = "http://www.myorg/repository",
     compliant = TRUE)
