### 2016-02-11 a7c5f2d
#88 improve request builders, handlers, add formatters for flexible customization of data provider config
