+++ R/SDMXREST21RequestBuilder-methods.R
@@ -31,92 +31,79 @@
 #'     compliant = TRUE)
 #'
 SDMXREST21RequestBuilder <- function(regUrl, repoUrl, compliant,
-                                   unsupportedResources = list(),
-                                   skipAgencyId = FALSE, forceAgencyId = FALSE){
+                                     unsupportedResources = list(),
+                                     skipAgencyId = FALSE, forceAgencyId = FALSE){
   
-  #function to handle request
-  serviceRequestHandler <- function(regUrl, repoUrl, agencyId, resource, resourceId, version = NULL,
-                                    flowRef, key = NULL, start = NULL, end = NULL,
-                                    compliant){    
-    
-    if(is.null(resource))
-      stop("Missing SDMX service resource")
-    
-    if(resource %in% unsupportedResources)
-      stop("Unsupported SDMX service resource for this provider")
-    
-    #wrap argument values
-    obj <- list(regUrl = regUrl, repoUrl = repoUrl,
-                agencyId = agencyId, resource = resource, resourceId = resourceId, version = version,
-                flowRef = flowRef, key = key, start = start, end = end)
-    
-    #REST resource handler
-    resourceHandler <- switch(resource,
+  #params formatter
+  formatter = list(
+    #dataflow
+    dataflow = function(obj){return(obj)},
+    #datastructure
+    datastructure = function(obj){ return(obj)},
+    #data
+    data = function(obj){return(obj)}
+  )
+  
+  #resource handler
+  handler <- list(
       
-      #'dataflow' resource (path="dataflow/{agencyID}/{resourceID}/{version}")
-      #-----------------------------------------------------------------------
-      "dataflow" = function(obj){
-        if(is.null(obj$resourceId)) obj$resourceId = "all"
-        if(is.null(obj$version)) obj$version = "latest"
-        req <- sprintf("%s/dataflow/%s/%s/%s",obj$regUrl, obj$agencyId, obj$resourceId, obj$version)        
-        return(req)
-      },
-                              
-      #'datastructure' resource (path="datastructure/{agencyID}/{resourceID}/{version})
-      #--------------------------------------------------------------------------------
-      "datastructure" = function(obj){    
-        if(is.null(obj$resourceId)) obj$resourceId = "all"
-        if(is.null(obj$version)) obj$version = "latest"       
-        req <- sprintf("%s/datastructure/%s/%s/%s",obj$regUrl, obj$agencyId, obj$resourceId, obj$version)
-        if(forceAgencyId) req <- paste(req, obj$agencyId, sep = "/")
-        req <- paste0(req, "?references=children") #TODO to see later to have arg for this
-        return(req)
-      },
-                              
-      #'data' resource (path="data/{flowRef}/{key}/{providerRef})
-      #----------------------------------------------------------
-      "data" = function(obj){
-        if(is.null(obj$flowRef)) stop("Missing flowRef value")
-        if(is.null(obj$key)) obj$key = "all"
-        req <- sprintf("%s/data/%s/%s",obj$repoUrl, obj$flowRef, obj$key)
-        if(skipAgencyId){
-          req <- paste0(req, "/")
-        }else{
-          req <- paste(req, ifelse(forceAgencyId, obj$agencyId, "all"), sep = "/")
-        }
-        
-        #DataQuery
-        #-> temporal extent (if any)
-        addParams = FALSE
-        if(!is.null(obj$start)){
-          req <- paste0(req, "?")
-          addParams = TRUE
-          req <- paste0(req, "startPeriod=", start)
-        }
-        if(!is.null(obj$end)){
-          if(!addParams){
[…]
