+++ R/SDMXServiceProvider-methods.R
@@ -62,230 +62,169 @@ SDMXServiceProvider <- function(agencyId, name,
 #'          \link{findSDMXServiceProvider} \link{readSDMX}
 #'
 setSDMXServiceProviders <- function(){ # nocov start
+    
+  #international data providers
+  #----------------------------
+
+  #ECB
+  ECB <- SDMXServiceProvider(
+    agencyId = "ECB", name = "European Central Bank",
+    builder = SDMXREST21RequestBuilder(
+      regUrl = "https://sdw-wsrest.ecb.europa.eu/service",
+      repoUrl = "https://sdw-wsrest.ecb.europa.eu/service",
+      compliant = TRUE)
+  )
+
+  #EUROSTAT
+  ESTAT <- SDMXServiceProvider( 
+    agencyId = "ESTAT", name = "Eurostat (Statistical office of the European Union)",
+    builder = SDMXREST21RequestBuilder(
+      regUrl = "http://ec.europa.eu/eurostat/SDMX/diss-web/rest",
+      repoUrl = "http://ec.europa.eu/eurostat/SDMX/diss-web/rest",
+      compliant = TRUE)
+  )
+
+  #OECD
+  OECD <- SDMXServiceProvider(
+    agencyId = "OECD", name = "Organisation for Economic Cooperation and Development ",
+    builder = SDMXDotStatRequestBuilder(
+      regUrl = "http://stats.oecd.org/restsdmx/sdmx.ashx",
+      repoUrl = "http://stats.oecd.org/restsdmx/sdmx.ashx")
+  )
+  
+  #UN-FAO
+  FAO <- SDMXServiceProvider(
+    agencyId = "FAO", name = "Food and Agriculture Organization of the United Nations",
+    builder = SDMXREST21RequestBuilder(
+      regUrl = "http://data.fao.org/sdmx/registry",
+      repoUrl = "http://data.fao.org/sdmx/repository",
+      compliant = FALSE,
+      unsupportedResources = list("dataflow"))
+  )
+  
+  #UN-ILO
+  ILO <- SDMXServiceProvider(
+    agencyId = "ILO", name = "International Labour Organization of the United Nations",
+    builder = SDMXREST21RequestBuilder(
+      regUrl = "http://www.ilo.org/ilostat/sdmx/ws/rest",
+      repoUrl = "http://www.ilo.org/ilostat/sdmx/ws/rest",
+      compliant = FALSE, skipAgencyId = TRUE,
+      unsupportedResources = list("dataflow"))                  
+  )
+  
+  #UIS (UNESCO)
+  UIS <- SDMXServiceProvider(
+    agencyId = "UIS", name = "UNESCO Institute of Statistics",
+    builder = SDMXDotStatRequestBuilder(
+      regUrl = "http://data.uis.unesco.org/RestSDMX/sdmx.ashx",
+      repoUrl = "http://data.uis.unesco.org/RestSDMX/sdmx.ashx")
+  )
+  
+  #national data providers
+  #-----------------------
+  
+  #ABS {Australia}
+  ABS <- SDMXServiceProvider(
+    agencyId = "ABS", name = "Australian Bureau of Statistics",
+    scale = "national", country = "AUS",
+    builder = SDMXDotStatRequestBuilder(
+      regUrl = "http://stat.abs.gov.au/restsdmx/sdmx.ashx",
+      repoUrl = "http://stat.abs.gov.au/restsdmx/sdmx.ashx", 
+      forceAgencyId = TRUE, unsupportedResources = list("dataflow"))
+  )
+  
+  #NBB {Belgium}
+  NBB <- SDMXServiceProvider(
+    agencyId = "NBB", name = "National Bank of Belgium",
[…]
