### 2016-02-11 b35b1db
#88 fix Rdoc and tests
