+++ R/SDMXCompactData-methods.R
@@ -133,7 +133,7 @@ as.data.frame.SDMXAllCompactData <- function(x, nsExpr, labels = FALSE, ...) {
 }
 
 
-as.data.frame.SDMXCompactData <- function(x, labels = FALSE, ...){
+as.data.frame.SDMXCompactData <- function(x, row.names, optional, labels = FALSE, ...){
   return(as.data.frame.SDMXAllCompactData(x, "compact", labels));
 }
 
