+++ R/SDMXREST20RequestBuilder-methods.R
@@ -27,8 +27,7 @@
 #'   #how to create a SDMXREST20RequestBuilder
 #'   requestBuilder <- SDMXREST20RequestBuilder(
 #'     regUrl = "http://www.myorg/registry",
-#'     repoUrl = "http://www.myorg/repository",
-#'     compliant = TRUE)
+#'     repoUrl = "http://www.myorg/repository", compliant = FALSE)
 #'
 SDMXREST20RequestBuilder <- function(regUrl, repoUrl, compliant,
                                      unsupportedResources = list(),
