### 2016-02-11 #15 @sckott
@eblondel Hows it going with changes in response to review?
