### 2016-02-11 b2ed512
#88 fix data enrichment with DSD + remove useless setAs
