+++ tests/testthat/test_Main_Helpers.R
@@ -18,7 +18,7 @@ context("SDMXHelpers")
 test_that("ECB - dataflow",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ECB", resource = "dataflow")
+  sdmx <- readSDMX(providerId = "ECB", resource = "dataflow")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataFlows")
   }
@@ -28,7 +28,7 @@ test_that("ECB - dataflow",{
 test_that("ECB - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ECB", resource = "datastructure", resourceId = "ECB_DD1")
+  sdmx <- readSDMX(providerId = "ECB", resource = "datastructure", resourceId = "ECB_DD1")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataStructureDefinition")
   }
@@ -38,7 +38,7 @@ test_that("ECB - datastructure",{
 test_that("ECB - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ECB", resource = "data",
+  sdmx <- readSDMX(providerId = "ECB", resource = "data",
                    flowRef = "DD", key = "M.SE.BSI_STF.RO.4F_N", key.mode = "SDMX",
                    start = 2010, end = 2010)
   if(!is.null(sdmx)){
@@ -53,7 +53,7 @@ test_that("ECB - data",{
 test_that("ESTAT - dataflow",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ESTAT", resource = "dataflow")
+  sdmx <- readSDMX(providerId = "ESTAT", resource = "dataflow")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataFlows")
   }
@@ -63,7 +63,7 @@ test_that("ESTAT - dataflow",{
 test_that("ESTAT - datastructure",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ESTAT", resource = "datastructure", resourceId = "DSD_nama_gdp_c")
+  sdmx <- readSDMX(providerId = "ESTAT", resource = "datastructure", resourceId = "DSD_nama_gdp_c")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataStructureDefinition")
   }
@@ -73,7 +73,7 @@ test_that("ESTAT - datastructure",{
 test_that("ESTAT - data",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "ESTAT", resource = "data",
+  sdmx <- readSDMX(providerId = "ESTAT", resource = "data",
                    flowRef = "cdh_e_fos", key = list(NULL, NULL, "PC", "FOS1", "BE"),
                    start = 2005, end = 2010)
   if(!is.null(sdmx)){
@@ -88,7 +88,7 @@ test_that("ESTAT - data",{
 test_that("OECD - dataflow",{
   testthat::skip_on_travis()
   testthat::skip_on_cran()
-  sdmx <- readSDMX(agencyId = "OECD", resource = "dataflow")
+  sdmx <- readSDMX(providerId = "OECD", resource = "dataflow")
   if(!is.null(sdmx)){
     expect_is(sdmx, "SDMXDataFlows")
   }
@@ -98,7 +98,7 @@ test_that("OECD - dataflow",{
 test_that("OECD - datastructure",{
   testthat::skip_on_travis()
[…]
