### 2016-02-12 2693cfc
#90 fix Rdoc and unit tests
