+++ tests/testthat/test_Data.R
@@ -33,14 +33,14 @@ test_that("DSD is properly fetched by readSDMX and associated to the dataset",{
 })
 
 test_that("DSD is properly fetched by readSDMX when there is no dsdRef (using flowRef)",{
-  data <- readSDMX(agencyId = "KNOEMA", resource = "data",
+  data <- readSDMX(providerId = "KNOEMA", resource = "data",
                    flowRef = "SADG2015", dsd = TRUE)
   expect_false(is.null(slot(data,"dsd")))
   expect_is(slot(data,"dsd"), "SDMXDataStructureDefinition")
 })
 
 test_that("Dataset is correctly enriched with labels using the DSD",{
-  sdmx.data <- readSDMX(agencyId = "UIS", resource = "data",
+  sdmx.data <- readSDMX(providerId = "UIS", resource = "data",
                         flowRef = "EDULIT_DS", key = list("OFST_1_CP", NULL),
                         start = "2000", end = "2015", dsd = TRUE)
   data <- as.data.frame(sdmx.data)
