+++ R/readSDMX.R
@@ -4,7 +4,7 @@
 #' @description \code{readSDMX} is the main function to use to read SDMX data
 #'
 #' @usage readSDMX(file, isURL,
-#'                 provider, agencyId, resource, resourceId, version,
+#'                 provider, providerId, agencyId, resource, resourceId, version,
 #'                 flowRef, key, key.mode, start, end, dsd, verbose)
 #'                 
 #' @param file path to SDMX-ML document that needs to be parsed
