### 2016-02-16 81b3403
refine README, add CI with opencpu cloud server
