### 2016-02-29 fc84d92
update SDMXRequestParams doc + #93 remove XML comments
