### 2016-03-13 820afc7
#94 readSDMX support for a 'validate' argument (default FALSE, skip SDMX strict compliance validation)
