+++ R/Class-SDMXConcept.R
@@ -66,10 +66,14 @@ setClass("SDMXConcept",
              fr = "description du concept"
            )
          ),
-         validity = function(object){  
-           #validation rules
-           if(is.na(object@id)) return(FALSE)
-           if(length(object@Name) == 0) return(FALSE)
+         validity = function(object){
+           
+           #eventual validation rules
+           if(.rsdmx.options$validate){
+            if(is.na(object@id)) return(FALSE)
+            if(length(object@Name) == 0) return(FALSE)
+           } 
+            
            return(TRUE);
          }
 )
