+++ R/Class-SDMXDimension.R
@@ -83,8 +83,12 @@ setClass("SDMXDimension",
           
          ),
          validity = function(object){
+           
            #eventual validation rules
-           #if(is.na(object@conceptRef)) return(FALSE)
+           if(.rsdmx.options$validate){
+            if(is.na(object@conceptRef)) return(FALSE)
+           }
+           
            return(TRUE);
          }
 )
