### 2016-03-13 7f80b21
#77 improve encoding of SDMX data.frame outputs
