### 2016-03-14 98c0223
#95 improve management of SDMX-ML namespaces
