+++ R/SDMXCodelist-methods.R
@@ -6,18 +6,18 @@
 #' SDMXCodelist(xmlObj)
 #' 
 #' @param xmlObj object of class "XMLInternalDocument derived from XML package
+#' @param namespaces object of class "data.frame" given the list of namespace URIs
 #' @return an object of class "SDMXCodelist"
 #' 
 #' @seealso \link{readSDMX}
 #'
-SDMXCodelist <- function(xmlObj){
+SDMXCodelist <- function(xmlObj, namespaces){
   
-  sdmxVersion <- version.SDMXSchema(xmlDoc(xmlObj))
-  VERSION.21 <- sdmxVersion == "2.1"
-  
-  namespaces <- namespaces.SDMX(xmlDoc(xmlObj))
   messageNs <- findNamespace(namespaces, "message")
   strNs <- findNamespace(namespaces, "structure")
+  
+  sdmxVersion <- version.SDMXSchema(xmlDoc(xmlObj), namespaces)
+  VERSION.21 <- sdmxVersion == "2.1"
   
   #attributes
   #=========
@@ -121,7 +121,7 @@ SDMXCodelist <- function(xmlObj){
   
   codes <- list()
   if(length(codesXML) > 0){
-    codes <- lapply(codesXML, function(x){ SDMXCode(x)})
+    codes <- lapply(codesXML, SDMXCode, namespaces)
   }
   
   #instantiate the object
