+++ R/SDMXDataFlow-methods.R
@@ -6,19 +6,19 @@
 #' SDMXDataFlow(xmlObj)
 #' 
 #' @param xmlObj object of class "XMLInternalDocument derived from XML package
+#' @param namespaces object of class "data.frame" given the list of namespace URIs
 #' @return an object of class "SDMXDataFlow"
 #' 
 #' @seealso \link{readSDMX}
 #'
-SDMXDataFlow <- function(xmlObj){
+SDMXDataFlow <- function(xmlObj, namespaces){
   
-  sdmxVersion <- version.SDMXSchema(xmlDoc(xmlObj))
-  VERSION.21 <- sdmxVersion == "2.1"
-  
-  namespaces <- namespaces.SDMX(xmlDoc(xmlObj))
   messageNs <- findNamespace(namespaces, "message")
   strNs <- findNamespace(namespaces, "structure")
   
+  sdmxVersion <- version.SDMXSchema(xmlDoc(xmlObj), namespaces)
+  VERSION.21 <- sdmxVersion == "2.1"
+
   #attributes
   #=========
   id = xmlGetAttr(xmlObj, "id")

[…]
