### 2016-03-14 8fcc502
downgrade XML revision, too early for OpenCPU rsdmx install
