### 2016-03-14 64edfc5
#95 more improvements & fix unit tests
