+++ tests/testthat/test_Codelists.R
@@ -9,8 +9,9 @@ context("SDMXCodelists")
 
 test_that("Codelists - 2.0",{
   file <- system.file("extdata", "SDMXCodelists_Example_2.0.xml", package = "rsdmx")
-  xmlObj <- xmlParse(file)  
-  codelists <- SDMXCodelists(xmlObj)
+  xmlObj <- xmlParse(file)
+  ns <- namespaces.SDMX(xmlObj)
+  codelists <- SDMXCodelists(xmlObj, ns)
   expect_is(codelists, "SDMXCodelists")
   expect_equal(length(codelists@codelists), 1L)
   
@@ -20,8 +21,9 @@ test_that("Codelists - 2.0",{
 
 test_that("Codelists - 2.1",{
   file <- system.file("extdata", "SDMXCodelists_Example_2.1.xml", package = "rsdmx")
-  xmlObj <- xmlParse(file)  
-  codelists <- SDMXCodelists(xmlObj)
+  xmlObj <- xmlParse(file)
+  ns <- namespaces.SDMX(xmlObj)
+  codelists <- SDMXCodelists(xmlObj, ns)
   expect_is(codelists, "SDMXCodelists")
   expect_equal(length(codelists@codelists), 1L)
   
