### 2016-03-14 1733110
#95 more fix on Rdoc...
