+++ R/SDMXTimeDimension-methods.R
@@ -3,7 +3,7 @@
 #' @aliases SDMXTimeDimension,SDMXTimeDimension-method
 #' 
 #' @usage
-#' SDMXTimeDimension(xmlObj)
+#' SDMXTimeDimension(xmlObj, namespaces)
 #' 
 #' @param xmlObj object of class "XMLInternalDocument derived from XML package
 #' @param namespaces object of class "data.frame" given the list of namespace URIs
