+++ R/SDMXDataFlows-methods.R
@@ -3,7 +3,7 @@
 #' @aliases SDMXDataFlows,SDMXDataFlows-method
 #' 
 #' @usage
-#' SDMXDataFlows(xmlObj)
+#' SDMXDataFlows(xmlObj, namespaces)
 #' 
 #' @param xmlObj object of class "XMLInternalDocument derived from XML package
 #' @param namespaces object of class "data.frame" given the list of namespace URIs
