+++ R/SDMXServiceProvider-methods.R
@@ -20,7 +20,7 @@
 #' @examples
 #'   #let's create a SDMXRESTRequestBuilder
 #'   #(assuming that "My Organization" implements SDMX REST web-services)
-#'   myBuilder <- SDMXRESTRequestBuilder(regUrl = "http://www.myorg.org/registry",
+#'   myBuilder <- SDMXREST20RequestBuilder(regUrl = "http://www.myorg.org/registry",
 #'                                       repoUrl = "http://www.myorg.org/repository",
 #'                                       compliant = TRUE)
 #'   
