+++ R/SDMXConcept-methods.R
@@ -3,7 +3,7 @@
 #' @aliases SDMXConcept,SDMXConcept-method
 #' 
 #' @usage
-#' SDMXConcept(xmlObj)
+#' SDMXConcept(xmlObj, namespaces)
 #' 
 #' @param xmlObj object of class "XMLInternalDocument derived from XML package
 #' @param namespaces object of class "data.frame" given the list of namespace URIs
