+++ R/SDMXServiceProvider-methods.R
@@ -87,23 +87,43 @@ setSDMXServiceProviders <- function(){ # nocov start
   #IMF
   IMF <- SDMXServiceProvider(
     agencyId = "IMF", name = "International Monetary Fund",
-    builder = SDMXDotStatRequestBuilder(
-      regUrl = "http://sdmxws.imf.org/SDMXRest/sdmx.ashx",
-      repoUrl = "http://sdmxws.imf.org/SDMXRest/sdmx.ashx")
+    builder = SDMXREST20RequestBuilder(
+      regUrl = "http://dataservices.imf.org/REST/SDMX_XML.svc",
+      repoUrl = "http://dataservices.imf.org/REST/SDMX_XML.svc",
+      compliant = FALSE)
   )
-  IMF@builder@formatter$dataflow <- function(obj){
-    obj@regUrl <- unlist(strsplit(obj@regUrl,"/sdmx.ashx"))[1]
-    return(obj)
+  IMF@builder@handler$dataflow <- function(obj){
+    req <- sprintf("%s/Dataflow/",obj@regUrl)
+    return(req)
+  }
+  IMF@builder@handler$datastructure <- function(obj){
+    req <- sprintf("%s/DataStructure/%s", obj@regUrl, obj@resourceId)
+    return(req)
   }
   IMF@builder@handler$data <- function(obj){
     if(is.null(obj@flowRef)) stop("Missing flowRef value")
-    req <- sprintf("%s/GetData?dataflow=%s", obj@repoUrl, obj@flowRef)
-    if(!is.null(obj@key)) req <- paste0(req, "&key=", obj@key)
+    if(is.null(obj@key)) obj@key = "."
+    
+    req <- sprintf("%s/CompactData/%s/%s",
+                   obj@repoUrl, obj@flowRef, obj@key)
     
     #DataQuery
     #-> temporal extent (if any)
-    if(!is.null(obj@start)) req <- paste0(req, "&startTime=", obj@start)
-    if(!is.null(obj@end)) req <- paste0(req, "&endTime=", obj@end) 
+    addParams = FALSE
+    if(!is.null(obj@start)){
+      req <- paste0(req, "?")
+      addParams = TRUE
+      req <- paste0(req, "startPeriod=", obj@start)
+    }
+    if(!is.null(obj@end)){
+      if(!addParams){
+        req <- paste0(req, "?")
+      }else{
+        req <- paste0(req, "&")
+      }
+      req <- paste0(req, "endPeriod=", obj@end) 
+    }
+    
     return(req)
   }
 
