### 2016-03-15 b9d31c5
some fixes in Rdoc before CRAN submission
