+++ man/addSDMXServiceProvider.Rd
@@ -16,7 +16,7 @@ function that allows configuring a new \link{SDMXServiceProvider}
 }
 \examples{
 #create a provider
-  myBuilder <- SDMXRESTRequestBuilder(regUrl = "http://www.myorg.org/registry",
+  myBuilder <- SDMXREST20RequestBuilder(regUrl = "http://www.myorg.org/registry",
                                       repoUrl = "http://www.myorg.org/repository",
                                       compliant = TRUE)
   myProvider <- SDMXServiceProvider(
