+++ man/isSoapRequestEnvelope.Rd
@@ -4,7 +4,7 @@
 \alias{isSoapRequestEnvelope}
 \title{isSoapRequestEnvelope}
 \usage{
-isSoapRequestEnvelope(xmlObj)
+isSoapRequestEnvelope(xmlObj, namespaces)
 }
 \arguments{
 \item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
