+++ R/SDMXGenericData-methods.R
@@ -233,7 +233,6 @@ as.data.frame.SDMXGenericData <- function(x, row.names=NULL, optional=FALSE,
   dataset <- do.call("rbind.fill", lapply(seriesXML, function(x){
     serie <- parseSerie(x)
   }))
-  colnames(dataset) <- serieNames
   dataset$obsValue <- as.numeric(dataset$obsValue)
   
   if(any(as.character(dataset$obsValue) == "NaN", na.rm = TRUE)){
