### 2016-03-16 83f73ee
minor fixes and prepare minor revision for CRAN
