### 2016-03-17 711747f
#96 provide workaround for handling EUROSTAT dataflows
