+++ man/rsdmx.Rd
@@ -17,8 +17,8 @@ currently focuses on the SDMX XML standard format (SDMX-ML).
    Package: \tab rsdmx\cr
    Type: \tab Package\cr
    Version
-   : \tab 0.5-3\cr
-   Date: \tab 2016-03-16\cr
+   : \tab 0.5-4\cr
+   Date: \tab 2016-03-17\cr
    License: \tab GPL(>=2.0)\cr
    LazyLoad: \tab yes\cr
  }
