+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rsdmx
-Version: 0.5-3
-Date: 2016-03-16
+Version: 0.5-4
+Date: 2016-03-17
 Title: Tools for Reading SDMX Data and Metadata
 Authors@R: c(
   person("Emmanuel", "Blondel", role = c("aut", "cre"), email = "emmanuel.blondel1@gmail.com"),
