### 2016-03-24 #15 @sckott
@eblondel Hows it going with changes in response to review?
