### 2016-03-24 #15 @eblondel
Hi @sckott, i believe most of the changes requested have yet been tackled in the past, however at now i've put this rsdmx submission in stand-by, because i want to go deeper in evaluating pros/cons, and look how policies might evolve at short-medium term. Counting with your understanding,
Best, Emmanuel
