### 2016-04-21 0e4923d
#96 fix 'ESTAT' datastructure query, #98 add 'WBG_WITS' provider
