### 2016-04-29 cdb9851
#99 fix minor bug in SDMXDataStructures as.data.frame method
