### 2016-04-30 426fa78
#99 fix small typo to make tests passing
