+++ man/rsdmx.Rd
@@ -18,7 +18,7 @@ currently focuses on the SDMX XML standard format (SDMX-ML).
    Type: \tab Package\cr
    Version
    : \tab 0.5-4\cr
-   Date: \tab 2016-04-30\cr
+   Date: \tab 2016-06-17\cr
    License: \tab GPL(>=2.0)\cr
    LazyLoad: \tab yes\cr
  }
