### 2016-07-21 6675ed1
#101 fix issue with multiple authority namespaces (loop until finding series)
