### 2016-07-21 a8aea1e
#101 fix small regression & test passing
