### 2016-07-22 9597002
#101 fix default lang assignment in case no lang xml attribute is found
