+++ R/SDMXDataFlow-methods.R
@@ -83,6 +83,8 @@ SDMXDataFlow <- function(xmlObj, namespaces){
     sapply(flowNamesXML,
            function(x){
              lang <- xmlGetAttr(x,"xml:lang")
+             if(is.null(lang)) lang <- xmlGetAttr(x,"lang")
+             if(is.null(lang)) lang <- "default"
              flowNames[[lang]] <- xmlValue(x)
            })
     flowNames <- as.list(flowNames)
@@ -113,6 +115,8 @@ SDMXDataFlow <- function(xmlObj, namespaces){
     sapply(flowDesXML,
            function(x){
              lang <- xmlGetAttr(x,"xml:lang")
+             if(is.null(lang)) lang <- xmlGetAttr(x,"lang")
+             if(is.null(lang)) lang <- "default"
              flowDescriptions[[lang]] <- xmlValue(x)
            })
     flowDescriptions <- as.list(flowDescriptions)
