### 2016-07-23 4122e14
#103 support for SDMX StructureSpecificTimeSeriesData document
