+++ R/Class-SDMXStructureSpecificTimeSeriesData.R
@@ -1,0 +1,22 @@
+#' @name SDMXStructureSpecificTimeSeriesData
+#' @docType class
+#' @aliases SDMXStructureSpecificTimeSeriesData-class
+#' 
+#' @title Class "SDMXStructureSpecificTimeSeriesData"
+#' @description A basic class to handle a SDMX-ML StructureSpecificTimeSeriesData data set
+#' 
+#' @section Warning:
+#' This class is not useful in itself, but all SDMX non-abstract classes will 
+#' encapsulate it as slot, when parsing an SDMX-ML document.
+#' 
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#' 
+setClass("SDMXStructureSpecificTimeSeriesData",
+         contains = "SDMXData",
+         representation(),
+         prototype = list(),
+         validity = function(object){
+           #eventual validation rules
+           return(TRUE);
+         }
+)
