+++ man/SDMXStructureSpecificTimeSeriesData.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2 (4.1.1): do not edit by hand
+% Please edit documentation in R/Class-SDMXStructureSpecificTimeSeriesData.R, R/SDMXStructureSpecificTimeSeriesData-methods.R
+\docType{class}
+\name{SDMXStructureSpecificTimeSeriesData}
+\alias{SDMXStructureSpecificTimeSeriesData}
+\alias{SDMXStructureSpecificTimeSeriesData,SDMXStructureSpecificTimeSeriesData-method}
+\alias{SDMXStructureSpecificTimeSeriesData-class}
+\title{Class "SDMXStructureSpecificTimeSeriesData"}
+\usage{
+SDMXStructureSpecificTimeSeriesData(xmlObj, namespaces)
+}
+\arguments{
+\item{xmlObj}{object of class "XMLInternalDocument derived from XML package}
+
+\item{namespaces}{object of class "data.frame" given the list of namespace URIs}
+}
+\value{
+an object of class "SDMXStructureSpecificTimeSeriesData"
+}
+\description{
+A basic class to handle a SDMX-ML StructureSpecificTimeSeriesData data set
+}
+\section{Warning}{
+
+This class is not useful in itself, but all SDMX non-abstract classes will
+encapsulate it as slot, when parsing an SDMX-ML document.
+}
+\author{
+Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
+}
+\seealso{
+\link{readSDMX}
+}
+
