### 2016-07-27 de09a9b
#91 remove unsupportedResources for ISTAT connector
