### 2016-08-02 e626d2d
#104 support DSD inheritance for SDMXDataFlows document
