### 2016-08-17 df0e23b
preparing CRAN submission
