### 2016-11-14 46d8f17
#108 fix typo in unit test
