+++ tests/testthat/test_Data.R
@@ -45,7 +45,7 @@ test_that("Dataset is correctly enriched with labels using the DSD",{
   data <- as.data.frame(sdmx.data)
   data.enriched <- as.data.frame(sdmx.data, labels = TRUE)
   expect_true(ncol(data.enriched) > ncol(data))
-  expect_true(all(data["CAI_IND"] == data.enriched["CAI_ID"]))
+  expect_true(all(data["CAI_IND"] == data.enriched["CAI_IND"]))
   expect_true(all(data["LOCATION"] == data.enriched["LOCATION"]))
   expect_true(all(data["TIME_FORMAT"] == data.enriched["TIME_FORMAT"]))
   expect_true(all(data["obsTime"] == data.enriched["obsTime"]))
