### 2016-11-14 6c6c7d6
prepare to CRAN, use CRAN canonical URLs with https
