### 2016-11-17 0de1314
Make readSDMX explicitly check for a 200 HTTP status.
