### 2016-11-21 b7b8cc5
update to current revision 0.5-7
