+++ R/rsdmx.R
@@ -16,8 +16,8 @@
 #'    Package: \tab rsdmx\cr
 #'    Type: \tab Package\cr
 #'    Version
-#'    : \tab 0.5-6\cr
-#'    Date: \tab 2016-11-14\cr
+#'    : \tab 0.5-7\cr
+#'    Date: \tab 2016-11-21\cr
 #'    License: \tab GPL(>=2.0)\cr
 #'    LazyLoad: \tab yes\cr
 #'  }
