+++ DESCRIPTION
@@ -1,10 +1,11 @@
 Package: rsdmx
-Version: 0.5-6
-Date: 2016-11-14
+Version: 0.5-7
+Date: 2016-11-21
 Title: Tools for Reading SDMX Data and Metadata
 Authors@R: c(
   person("Emmanuel", "Blondel", role = c("aut", "cre"), email = "emmanuel.blondel1@gmail.com"),
-  person("Matthieu", "Stigler", role = c("ctb")))
+  person("Matthieu", "Stigler", role = c("ctb")),
+  person("Eric", "Persson", role = c("ctb")))
 Maintainer: Emmanuel Blondel <emmanuel.blondel1@gmail.com>
 Depends: R (>= 2.15)
 Imports: methods, XML (>= 3.98.1.3), RCurl, plyr, utils
