### 2016-11-23 b3e421b
#112 fix readSDMX with provider object arg + improve tests
