### 2016-11-24 34669e6
#114 add tentative connector Widukind SDMX service
