### 2016-11-25 819a3f3
#115 add/test NOMIS SDMX Service provider
