### 2016-12-21 11a2df8
#116 add saveSDMX & support of RData file in readSDMX
