+++ NAMESPACE
@@ -93,7 +93,8 @@ export(
   findSDMXServiceProvider,
   setSDMXServiceProviders,
   getSDMXServiceProviders,
-	readSDMX
+	readSDMX,
+  saveSDMX
 )
 
 exportMethods(
