### 2016-12-21 9e22d88
#116 add section to README, vignette
