+++ R/Class-SDMXDotStatRequestBuilder.R
@@ -8,6 +8,8 @@
 #' 
 #' @slot regUrl an object of class "character" giving the base Url of the SDMX service registry
 #' @slot repoUrl an object of class "character" giving the base Url of the SDMX service repository
+#' @slot accessKey an object of class "character" indicating the name of request parameter for which
+#'       an authentication or subscription user key/token has to be provided to perform requests 
 #' @slot compliant an object of class "logical" indicating if the request builder is somehow compliant with a service specification
 #'
 #' @section Warning:
