+++ R/saveSDMX.R
@@ -5,7 +5,7 @@
 #'
 #' @usage saveSDMX(sdmxObj, file)
 #'                 
-#' @param sdmxObj an object of class \code{\link{SDMX-Class}} to save
+#' @param sdmxObj an object of class \code{\link{SDMX-class}} to save
 #' @param file a connection or the name of the file where the R object is saved to
 #' 
 #' @examples             
