+++ R/Class-SDMXREST21RequestBuilder.R
@@ -7,6 +7,8 @@
 #' 
 #' @slot regUrl an object of class "character" giving the base Url of the SDMX service registry
 #' @slot repoUrl an object of class "character" giving the base Url of the SDMX service repository
+#' @slot accessKey an object of class "character" indicating the name of request parameter for which
+#'       an authentication or subscription user key/token has to be provided to perform requests 
 #' @slot compliant an object of class "logical" indicating if the web-service is compliant with the SDMX REST web-service specifications
 #'
 #' @section Warning:
