### 2017-01-23 b9f50fb
#117 fix unit tests
