### 2017-01-24 b6c883b
#119 fix minor issue in parsing generic data (when only 1 data dimension)
