### 2017-01-28 964cddf
#120 fix parsing values with "," decimal
