### 2017-01-29 542db79
#122 allow DSD labels association from codelists only (when DSD is only compound of codelists)
