### 2017-01-29 294b5a9
#121 inherit DSD in case of embedded codelist or DSD, for data enrichment with labels
