+++ R/readSDMX.R
@@ -350,54 +350,71 @@ readSDMX <- function(file = NULL, isURL = TRUE, isRData = FALSE,
     }
   }
   
-  #attempt to get DSD in case of helper method
-  if(buildRequest && resource %in% c("data","dataflow") && dsd){
+  #attempt to get DSD
+  embeddedDSD <- FALSE
+  if(is(obj, "SDMXData")){
+    strTypeObj <- SDMXStructureType(obj@xmlObj, ns, NULL)
+    if(strTypeObj@subtype %in% c("CodelistsType", "DataStructureDefinitionsType")){
+      dsd <- TRUE
+      embeddedDSD <- TRUE
+    }
+  }
+  
+  if(dsd){
+    dsdObj <- NULL
     
-    if(resource == "data" && providerId %in% c("ESTAT", "ISTAT", "WBG_WITS", "UIS2")){
-      if(verbose) message("-> Attempt to fetch DSD ref from dataflow description")
-      flow <- readSDMX(providerId = providerId, providerKey = providerKey, resource = "dataflow",
-                      resourceId = flowRef, verbose = TRUE)
-      dsdRef <- slot(slot(flow, "dataflows")[[1]],"dsdRef")
-      rm(flow)
-    }else{
-      dsdRef <- NULL
-      if(resource == "data"){
-        dsdRef <- slot(obj, "dsdRef")
-      }else if(resource=="dataflow"){
-        dsdRef <- lapply(slot(obj,"dataflows"), slot,"dsdRef")
-      }
-      if(!is.null(dsdRef)){
-        if(verbose) message(paste0("-> DSD ref identified in dataset = '", dsdRef, "'"))
-        if(verbose) message("-> Attempt to fetch & bind DSD to dataset")
-      }else{
-        dsdRef <- flowRef
-        if(verbose) message("-> No DSD ref associated to dataset")
-        if(verbose) message("-> Attempt to fetch & bind DSD to dataset using 'flowRef'")
-      }
+    #in case codelist or DSD are embedded with data
+    if(embeddedDSD){
+      dsdObj <- SDMXDataStructureDefinition(obj@xmlObj, ns)
+      slot(obj, "dsd") <- dsdObj
     }
     
-    #fetch DSD
-    dsdObj <- NULL
-    if(resource == "data"){
-      dsdObj <- readSDMX(providerId = providerId, providerKey = providerKey,
-                         resource = "datastructure", resourceId = dsdRef, verbose = verbose)
-      if(is.null(dsdObj)){
-        if(verbose) message(sprintf("-> Impossible to fetch DSD for dataset %s", flowRef))
+    #using helpers strategy (with a resource parameter)
+    if(buildRequest && resource %in% c("data","dataflow")){
+      if(resource == "data" && providerId %in% c("ESTAT", "ISTAT", "WBG_WITS", "UIS2")){
+        if(verbose) message("-> Attempt to fetch DSD ref from dataflow description")
+        flow <- readSDMX(providerId = providerId, providerKey = providerKey, resource = "dataflow",
+                         resourceId = flowRef, verbose = TRUE)
+        dsdRef <- slot(slot(flow, "dataflows")[[1]],"dsdRef")
+        rm(flow)
       }else{
-        if(verbose) message("-> DSD fetched and associated to dataset!")
-        slot(obj, "dsd") <- dsdObj
-      }
-    }else if(resource == "dataflow"){
-      dsdObj <- lapply(1:length(dsdRef), function(x){
-        flowDsd <- readSDMX(providerId = providerId, providerKey = providerKey,
-                            resource = "datastructure", resourceId = dsdRef[[x]], verbose = verbose)
-        if(is.null(flowDsd)){
-          if(verbose) message(sprintf("-> Impossible to fetch DSD for dataflow %s",resourceId))
[…]
