### 2017-01-29 d9fcfc3
#121 allow NULL subtype in SDMXStructureType for backward compatibility
