### 2017-01-29 5967467
#121 fix SDMXStructureType parser
