### 2017-01-30 bb0bc66
#123 replacing invalid XML entities prior to XML parsing
