### 2017-02-20 e4faf8b
deactivate eurostat bulk download example + hide results in vignette
