+++ vignettes/quickstart.Rmd
@@ -41,7 +41,7 @@ Let's see then how to use ``rsdmx``!
 
 ``rsdmx`` can be installed from CRAN or from its development repository hosted in Github. For the latter, you will need the ``devtools`` package and run:
 
-```{r, eval=FALSE}
+```{r, eval=FALSE, results="hide"}
 devtools::install_github("opensdmx/rsdmx")
 ```
 
@@ -49,7 +49,7 @@ devtools::install_github("opensdmx/rsdmx")
 
 To load rsdmx in R, do the following:
 
-```{r}
+```{r,results="hide"}
 library(rsdmx)
 ```
 
@@ -63,7 +63,7 @@ This section will introduce you on how to read SDMX *dataset* documents, either 
 
 The following code snipet shows you how to read a dataset from a remote data source, taking as example the [OECD StatExtracts portal](http://stats.oecd.org): [http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011](http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011)
 
-```{r}
+```{r,results="hide"}
 myUrl <- "http://stats.oecd.org/restsdmx/sdmx.ashx/GetData/MIG/TOT../OECD?startTime=2000&endTime=2011"
 dataset <- readSDMX(myUrl)
 stats <- as.data.frame(dataset)
@@ -77,7 +77,7 @@ The online rsdmx documentation also provides a list of data providers, either fr
 
 Now, the service providers above mentioned are known by ``rsdmx`` which let users using ``readSDMX`` with the helper parameters. The list of service providers can be retrieved doing:
 
-```{r}
+```{r,results="hide"}
 providers <- getSDMXServiceProviders();
 as.data.frame(providers)
 ```
@@ -86,7 +86,7 @@ Note it is also possible to add an SDMX service provider at runtime. For registe
 
 Let's see how it would look like for querying an ``OECD`` datasource:
 
-```{r, message = FALSE}
+```{r, message = FALSE,results="hide"}
 sdmx <- readSDMX(providerId = "OECD", resource = "data", flowRef = "MIG",
                 key = list("TOT", NULL, NULL), start = 2010, end = 2011)
 df <- as.data.frame(sdmx)
@@ -102,7 +102,7 @@ To do so, you only need to append ``dsd = TRUE`` (default value is ``FALSE``),
 to the previous request, and specify ``labels = TRUE`` when calling ``as.data.frame``,
 as follows:
 
-```{r, message = FALSE}
+```{r, message = FALSE,results="hide"}
 sdmx <- readSDMX(providerId = "OECD", resource = "data", flowRef = "MIG",
                 key = list("TOT", NULL, NULL), start = 2010, end = 2011,
                 dsd = TRUE)
@@ -119,7 +119,7 @@ Note that in case you are reading SDMX-ML documents with the native approach (wi
 URLs), instead of the embedded providers, it is also possible to associate a DSD
 to a dataset by using the function ``setDSD``. Let's try how it works:
 
-```{r, message = FALSE}
+```{r, message = FALSE,results="hide"}
 #data without DSD
 sdmx.data <- readSDMX(providerId = "OECD", resource = "data", flowRef = "MIG",
                 key = list("TOT", NULL, NULL), start = 2010, end = 2011)
@@ -136,7 +136,7 @@ sdmx.data <- setDSD(sdmx.data, sdmx.dsd)
 
[…]
