### 2017-02-20 d2c7fdc
deactivate vignette code
