### 2017-02-21 61ac003
change ILO example request in README
