### 2017-05-01 bf52074
Try fixing .travis.yml config
