### 2017-05-02 c3c6ce0
#126 fix issue in DSD reader - codelist name can be null
