### 2017-05-03 edbeed3
Update .travis.yml
