### 2017-06-14 2561d58
deactivate some live test
