### 2017-07-26 629e847
update date for release in CRAN
