### 2017-08-09 c78091d
#132 fix parsing of MessageGroup
