### 2017-08-09 bf0a07d
fix typo in DESCRIPTION
