+++ R/SDMXMessageGroup-methods.R
@@ -47,8 +47,9 @@ class.SDMXMessageGroup <- function(xmlObj){
     seriesKeyXML <- getNodeSet(xmlObj, "//ns:SeriesKey", c(ns = authorityNs$uri)) 
   }else{
     if(nrow(nsDefs.df) > 0){
-      seriesKeyXML <- getNodeSet(xmlObj, "//ns:SeriesKey",
-                                 c(ns = nsDefs.df[regexpr("generic$", nsDefs.df$uri)>0,"uri"]))
+      serieNs <- nsDefs.df[regexpr("generic$", nsDefs.df$uri)>0,"uri"]
+      if(length(serieNs)==0) serieNs <- nsDefs.df[1,"uri"]
+      seriesKeyXML <- getNodeSet(xmlObj, "//ns:SeriesKey", c(ns = serieNs))
     }else{    
       stop("Unsupported XML parser for empty target XML namespace")
     }
