### 2017-10-01 686145b
#133 fix generic data parser when no time
